==> Scripts_analyses/01_script_2017-09-04_randomization_for_respirometry.Rmd <==
---
title: "Randomizing groups"
output: html_document
---

### Day 2 cohort; 2017-08-25 collected   

```{r day 2}
day2<-read.csv("../Data/2017-09-04_day2cohort_randomization.csv")
day2

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day2$rando_assign<-sample(v,length(day2[,1]),replace=TRUE)
day2

ggplot(day2,aes(y=mass,x=rando_assign))+geom_boxplot()

ggplot(day2,aes(x=mass,colour=rando_assign,fill=rando_assign))+geom_density(alpha=.5,position="identity")#+geom_histogram(binwidth=1,alpha=.5)


summary(as.factor(day2$rando_assign))
summary(aov(mass~rando_assign,data=day2))
#write.csv(day2,"../Data/2017-09-04_day2cohort_randomized.csv")
day2$tape<-as.character(day2$tape)
##assign names cohort day, tape color, ind_id
day2$uniqueID<-paste(day2$cohort_day,substr(day2$tape,1,1),day2$Ind_ID,sep="")


gc<-subset(day2,day2$rando_assign=="genetic control")
dim(gc)
write.csv(gc,"../Data/randomizing/2017-09-04_day2cohort_genetic_control.csv")


RT<-subset(day2,day2$rando_assign=="RT eclosion")
write.csv(RT,"../Data/randomizing/2017-09-04_day2cohort_RT_eclosion.csv")


SO<-subset(day2,day2$rando_assign=="Simulated Overwintering")
write.csv(SO,"../Data/randomizing/2017-09-04_day2cohort_simulated_overwintering.csv")
```

### Day 3 cohort; 2017-08-26 collected   

```{r day 3}
day3<-read.csv("../Data/randomizing/2017-09-07_day3cohort_randomization.csv")
day3

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day3$rando_assign<-sample(v,length(day3[,1]),replace=TRUE)


ggplot(day3,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

ggplot(day3,aes(x=mass_day10,colour=rando_assign,fill=rando_assign))+geom_density(alpha=.5,position="identity")#+geom_histogram(binwidth=1,alpha=.5)


summary(as.factor(day3$rando_assign))
summary(aov(mass_day10~rando_assign,data=day3))
#write.csv(day2,"../Data/2017-09-04_day2cohort_randomized.csv")
day3$tape<-as.character(day3$tape)
##assign names cohort day, tape color, ind_id
day3$uniqueID<-paste(day3$cohort_day,substr(day3$tape,1,1),day3$Ind_ID,sep="")


gc<-subset(day3,day3$rando_assign=="genetic control")
dim(gc)
write.csv(gc,"../Data/randomizing/2017-09-07_day3cohort_genetic_control.csv")


RT<-subset(day3,day3$rando_assign=="RT eclosion")
write.csv(RT,"../Data/randomizing/2017-09-07_day3cohort_RT_eclosion.csv")


SO<-subset(day3,day3$rando_assign=="Simulated Overwintering")
write.csv(SO,"../Data/randomizing/2017-09-07_day3cohort_simulated_overwintering.csv")
```


### Day 4 cohort; 2017-08-27 collected   

```{r day 4}
day4<-read.csv("../Data/randomizing/2017-09-10_day4cohort_randomization.csv")
day4

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day4$rando_assign<-sample(v,length(day4[,1]),replace=TRUE)


ggplot(day4,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

#ggplot(day4,aes(x=mass_day10,colour=rando_assign,fill=rando_assign))+geom_density(alpha=.5,position="identity")#+geom_histogram(binwidth=1,alpha=.5)


summary(as.factor(day4$rando_assign))
summary(aov(mass_day10~rando_assign,data=day4))
#write.csv(day4,"../Data/2017-09-10_day4cohort_randomized.csv")
day4$tape<-as.character(day4$tape)
##assign names cohort day, tape color, ind_id
day4$uniqueID<-paste(day4$cohort_day,substr(day4$tape,1,1),day4$Ind_ID,sep="")


gc<-subset(day4,day4$rando_assign=="genetic control")
dim(gc)
write.csv(gc,"../Data/randomizing/2017-09-10_day4cohort_genetic_control.csv")


RT<-subset(day4,day4$rando_assign=="RT eclosion")
write.csv(RT,"../Data/randomizing/2017-09-10_day4cohort_RT_eclosion.csv")


SO<-subset(day4,day4$rando_assign=="Simulated Overwintering")
write.csv(SO,"../Data/randomizing/2017-09-10_day4cohort_simulated_overwintering.csv")
```

### Day 5 cohort; 2017-08-28 collected   

```{r day 5}
day5<-read.csv("../Data/randomizing/2017-09-11_day5cohort_randomization.csv")
day5

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day5$rando_assign<-sample(v,length(day5[,1]),replace=TRUE)


ggplot(day5,aes(y=mass_day5,x=rando_assign))+geom_boxplot()

#ggplot(day4,aes(x=mass_day10,colour=rando_assign,fill=rando_assign))+geom_density(alpha=.5,position="identity")#+geom_histogram(binwidth=1,alpha=.5)


summary(as.factor(day5$rando_assign))
summary(aov(mass_day10~rando_assign,data=day5))
#write.csv(day5,"../Data/randomizing/2017-09-11_day5cohort_randomized.csv")
day5$tape<-as.character(day5$tape)
##assign names cohort day, tape color, ind_id
day5$uniqueID<-paste(day5$cohort_day,substr(day5$tape,1,1),day5$Ind_ID,sep="")


gc<-subset(day5,day5$rando_assign=="genetic control")
dim(gc)
write.csv(gc,"../Data/randomizing/2017-09-11_day5cohort_genetic_control.csv")


RT<-subset(day5,day5$rando_assign=="RT eclosion")
write.csv(RT,"../Data/randomizing/2017-09-11_day5cohort_RT_eclosion.csv")


SO<-subset(day5,day5$rando_assign=="Simulated Overwintering")
write.csv(SO,"../Data/randomizing/2017-09-11_day5cohort_simulated_overwintering.csv")
```

### Day 10 cohort; 2017-09-02 collected   

```{r day 10}
day10<-read.csv("../Data/randomizing/2017-09-15_day10cohort_randomization.csv")
day10

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day10$rando_assign<-sample(v,length(day10[,1]),replace=TRUE)


ggplot(day10,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

#ggplot(day4,aes(x=mass_day10,colour=rando_assign,fill=rando_assign))+geom_density(alpha=.5,position="identity")#+geom_histogram(binwidth=1,alpha=.5)


summary(as.factor(day10$rando_assign))
summary(aov(mass_day10~rando_assign,data=day10))
#write.csv(day10,"../Data/randomizing/2017-09-15_day10cohort_randomized.csv")
day10$tape<-as.character(day10$tape)
##assign names cohort day, tape color, ind_id
day10$uniqueID<-paste(day10$cohort_day,substr(day10$tape,1,1),day10$Ind_ID,sep="")

day10$rando_assign<-as.factor(day10$rando_assign)
gc<-subset(day10,rando_assign=="genetic control")
dim(gc)
write.csv(gc,"../Data/randomizing/2017-09-15_day10cohort_genetic_control.csv")


RT<-subset(day10,rando_assign=="RT eclosion")
write.csv(RT,"../Data/randomizing/2017-09-15_day10cohort_RT_eclosion.csv")


SO<-subset(day10,day10$rando_assign=="Simulated Overwintering")
write.csv(SO,"../Data/randomizing/2017-09-15_day10cohort_simulated_overwintering.csv")
```


### Day 11 cohort; 2017-09-03 collected
```{r day 11}
day11<-read.csv("../Data/randomizing/2017-09-15_day11cohort_randomization.csv")
day11

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day11$rando_assign<-sample(v,length(day11[,1]),replace=TRUE)


ggplot(day11,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day11$rando_assign))
summary(aov(mass_day10~rando_assign,data=day11))
day11$tape<-as.character(day11$tape)
##assign names cohort day, tape color, ind_id
day11$uniqueID<-paste(day11$cohort_day,substr(day11$tape,1,1),day11$Ind_ID,sep="")
#write.csv(day11,"../Data/randomizing/2017-09-15_day11cohort_randomized.csv")

gc<-subset(day11,rando_assign=="genetic control")
dim(gc)
write.csv(gc,"../Data/randomizing/2017-09-15_day11cohort_genetic_control.csv")


RT<-subset(day11,rando_assign=="RT eclosion")
write.csv(RT,"../Data/randomizing/2017-09-15_day11cohort_RT_eclosion.csv")


SO<-subset(day11,rando_assign=="Simulated Overwintering")
write.csv(SO,"../Data/randomizing/2017-09-15_day11cohort_simulated_overwintering.csv")
```


### Day 12 cohort; 2017-09-04 collected
```{r day 12}
day12<-read.csv("../Data/randomizing/2017-09-15_day12cohort_randomization.csv")
day12

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day12$rando_assign<-sample(v,length(day12[,1]),replace=TRUE)


ggplot(day12,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day12$rando_assign))
summary(aov(mass_day10~rando_assign,data=day12))
day12$tape<-as.character(day12$tape)
##assign names cohort day, tape color, ind_id
day12$uniqueID<-paste(day12$cohort_day,substr(day12$tape,1,1),day12$Ind_ID,sep="")
#write.csv(day12,"../Data/randomizing/2017-09-15_day12cohort_randomized.csv")

gc<-subset(day12,rando_assign=="genetic control")
dim(gc)
write.csv(gc,"../Data/randomizing/2017-09-15_day12cohort_genetic_control.csv")


RT<-subset(day12,rando_assign=="RT eclosion")
write.csv(RT,"../Data/randomizing/2017-09-15_day12cohort_RT_eclosion.csv")


SO<-subset(day12,rando_assign=="Simulated Overwintering")
write.csv(SO,"../Data/randomizing/2017-09-15_day12cohort_simulated_overwintering.csv")
```

### Day 13 cohort; 2017-09-05 collected
```{r day 14}
day13<-read.csv("../Data/randomizing/2017-09-18-day13cohort_randomization.csv")
day13

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day13$rando_assign<-sample(v,length(day13[,1]),replace=TRUE)


ggplot(day13,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day13$rando_assign))
summary(aov(mass_day10~rando_assign,data=day13))
day13$tape<-as.character(day13$tape)
##assign names cohort day, tape color, ind_id
day13$uniqueID<-paste(day13$cohort_day,substr(day13$tape,1,1),day13$Ind_ID,sep="")
#write.csv(day13,"../Data/randomizing/2017-09-18_day13cohort_randomized.csv")

gc<-subset(day13,rando_assign=="genetic control")
dim(gc)
write.csv(gc,"../Data/randomizing/2017-09-18_day13cohort_genetic_control.csv")


RT<-subset(day13,rando_assign=="RT eclosion")
write.csv(RT,"../Data/randomizing/2017-09-18_day13cohort_RT_eclosion.csv")


SO<-subset(day13,rando_assign=="Simulated Overwintering")
write.csv(SO,"../Data/randomizing/2017-09-18_day13cohort_simulated_overwintering.csv")
```

### Day 14 cohort; 2017-09-06 collected
```{r day 14}
day14<-read.csv("../Data/randomizing/2017-09-18-day14cohort_randomization.csv")
day14

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day14$rando_assign<-sample(v,length(day14[,1]),replace=TRUE)


ggplot(day14,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day14$rando_assign))
summary(aov(mass_day10~rando_assign,data=day14))
day14$tape<-as.character(day14$tape)
##assign names cohort day, tape color, ind_id
day14$uniqueID<-paste(day14$cohort_day,substr(day14$tape,1,1),day14$Ind_ID,sep="")
#write.csv(day14,"../Data/randomizing/2017-09-18_day14cohort_randomized.csv")

gc<-subset(day14,rando_assign=="genetic control")
dim(gc)
write.csv(gc,"../Data/randomizing/2017-09-18_day14cohort_genetic_control.csv")


RT<-subset(day14,rando_assign=="RT eclosion")
write.csv(RT,"../Data/randomizing/2017-09-18_day14cohort_RT_eclosion.csv")


SO<-subset(day14,rando_assign=="Simulated Overwintering")
write.csv(SO,"../Data/randomizing/2017-09-18_day14cohort_simulated_overwintering.csv")
```


# creating randomized wells 

```{r}
letters<-rep(c("A","B","C","D","E","F","G","H"),12)

num<-sort(rep(seq(1,12,1),8))

wells<-paste(letters,num,sep="")
```


### Day 16 cohort; 2017-09-08 collected
```{r day 16}
day16<-read.csv("../Data/randomizing/2017-09-22-day16cohort_randomization.csv")
day16

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day16$rando_assign<-sample(v,length(day16[,1]),replace=TRUE)


ggplot(day16,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day16$rando_assign))
summary(aov(mass_day10~rando_assign,data=day16))
day16$tape<-as.character(day16$tape)
##assign names cohort day, tape color, ind_id
day16$uniqueID<-paste(day16$cohort_day,substr(day16$tape,1,1),day16$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
#sample(wells,length(day16$Ind_ID),replace=TRUE)
#day16$wellID<-ifelse(day16$rando_assign!="genetic control",sample(wells,length(day16$Ind_ID),replace=TRUE),"")

day16$wellID<-ifelse(day16$rando_assign=="RT eclosion",sample(wells,length(day16$Ind_ID),replace=TRUE),ifelse(day16$rando_assign=="Simulated Overwintering",sample(wells,length(day16$Ind_ID),replace=TRUE),""))
day16$wellID

#write.csv(day16,"../Data/randomizing/2017-09-22_day16cohort_randomized.csv")

```

### Day 18 cohort; 2017-09-10 collected

```{r day 18}
day18<-read.csv("../Data/randomizing/2017-09-22-day18cohort_randomization.csv")
day18

v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day18$rando_assign<-sample(v,length(day18[,1]),replace=TRUE)


ggplot(day18,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day18$rando_assign))
summary(aov(mass_day10~rando_assign,data=day18))
day18$tape<-as.character(day18$tape)
##assign names cohort day, tape color, ind_id
day18$uniqueID<-paste(day18$cohort_day,substr(day18$tape,1,1),day18$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
#sample(wells,length(day18$Ind_ID),replace=TRUE)
#day18$wellID<-ifelse(day18$rando_assign!="genetic control",sample(wells,length(day18$Ind_ID),replace=TRUE),"")

day18$wellID<-ifelse(day18$rando_assign=="RT eclosion",sample(wells,length(day18$Ind_ID),replace=TRUE),ifelse(day18$rando_assign=="Simulated Overwintering",sample(wells,length(day18$Ind_ID),replace=TRUE),""))
day18$wellID

#write.csv(day18,"../Data/randomizing/2017-09-22_day18cohort_randomized.csv")

```


### Day 19 cohort; 2017-09-11 collected
```{r day 19}
day19<-read.csv("../Data/randomizing/2017-09-22-day19cohort_randomization.csv")
day19<-day19[-67:-79,]
day19
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day19$rando_assign<-sample(v,length(day19[,1]),replace=TRUE)


ggplot(day19,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day19$rando_assign))
summary(aov(mass_day10~rando_assign,data=day19))
day19$tape<-as.character(day19$tape)
##assign names cohort day, tape color, ind_id
day19$uniqueID<-paste(day19$cohort_day,substr(day19$tape,1,1),day19$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
#sample(wells,length(day19$Ind_ID),replace=TRUE)
#day19$wellID<-ifelse(day19$rando_assign!="genetic control",sample(wells,length(day19$Ind_ID),replace=TRUE),"")

day19$wellID<-ifelse(day19$rando_assign=="RT eclosion",sample(wells,length(day19$Ind_ID),replace=TRUE),ifelse(day19$rando_assign=="Simulated Overwintering",sample(wells,length(day19$Ind_ID),replace=TRUE),""))
day19$wellID

#write.csv(day19,"../Data/randomizing/2017-09-22_day19cohort_randomized.csv")

```

### Day 20 cohort; 2017-09-12 collected  

```{r day 20}
day20<-read.csv("../Data/randomizing/2017-09-25-day20cohort_randomization.csv")
day20
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day20$rando_assign<-sample(v,length(day20[,1]),replace=TRUE)


ggplot(day20,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day20$rando_assign))
summary(aov(mass_day10~rando_assign,data=day20))
day20$tape<-as.character(day20$tape)
##assign names cohort day, tape color, ind_id
day20$uniqueID<-paste(day20$cohort_day,substr(day20$tape,1,1),day20$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day20$wellID<-ifelse(day20$rando_assign=="RT eclosion",sample(wells,length(day20$Ind_ID),replace=FALSE),ifelse(day20$rando_assign=="Simulated Overwintering",sample(wells,length(day20$Ind_ID),replace=FALSE),""))
day20$wellID

#write.csv(day20,"../Data/randomizing/2017-09-25_day20cohort_randomized.csv")

```

## haws 

### 2017-09-23 cohort day 2  

```{r}
day2<-read.csv("../Data/randomizing/2017-10-05_day2_haw_cohort_randomization.csv")
day2
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day2$rando_assign<-sample(v,length(day2[,1]),replace=TRUE)


ggplot(day2,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day2$rando_assign))
summary(aov(mass_day10~rando_assign,data=day2))
day2$tape<-as.character(day2$tape)
##assign names cohort day, tape color, ind_id
day2$uniqueID<-paste(day2$cohort_day,substr(day2$tape,1,1),day2$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day2$wellID<-ifelse(day2$rando_assign=="RT eclosion",sample(wells,length(day2$Ind_ID),replace=FALSE),ifelse(day2$rando_assign=="Simulated Overwintering",sample(wells,length(day2$Ind_ID),replace=FALSE),""))
day2$wellID

#write.csv(day2,"../Data/randomizing/2017-10-05_day2cohort_randomized.csv")

```

### 2017-09-24 cohort day 3  

```{r}
day3<-read.csv("../Data/randomizing/2017-10-05_day3_haw_cohort_randomization.csv")
day3
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day3$rando_assign<-sample(v,length(day3[,1]),replace=TRUE)


ggplot(day3,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day3$rando_assign))
summary(aov(mass_day10~rando_assign,data=day3))
day3$tape<-as.character(day3$tape)
##assign names cohort day, tape color, ind_id
day3$uniqueID<-paste(day3$cohort_day,substr(day3$tape,1,1),day3$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day3$wellID<-ifelse(day3$rando_assign=="RT eclosion",sample(wells,length(day3$Ind_ID),replace=FALSE),ifelse(day3$rando_assign=="Simulated Overwintering",sample(wells,length(day3$Ind_ID),replace=FALSE),""))
day3$wellID

#write.csv(day3,"../Data/randomizing/2017-10-05_day3_haw_cohort_randomized.csv")

```


### 2017-09-25 cohort day 4  

```{r}
day4<-read.csv("../Data/randomizing/2017-10-05_day4_haw_cohort_randomization.csv")
day4
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day4$rando_assign<-sample(v,length(day4[,1]),replace=TRUE)


ggplot(day4,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day4$rando_assign))
summary(aov(mass_day10~rando_assign,data=day4))
day4$tape<-as.character(day4$tape)
##assign names cohort day, tape color, ind_id
day4$uniqueID<-paste(day4$cohort_day,substr(day4$tape,1,1),day4$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day4$wellID<-ifelse(day4$rando_assign=="RT eclosion",sample(wells,length(day4$Ind_ID),replace=FALSE),ifelse(day4$rando_assign=="Simulated Overwintering",sample(wells,length(day4$Ind_ID),replace=FALSE),""))
day4$wellID

#write.csv(day4,"../Data/randomizing/2017-10-05_day4_haw_cohort_randomized.csv")

```


### 2017-09-26 cohort day 5  

```{r day 5}
day5<-read.csv("../Data/randomizing/2017-10-10_day5_haw_cohort_randomization.csv")
day5
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day5$rando_assign<-sample(v,length(day5[,1]),replace=TRUE)


ggplot(day5,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day5$rando_assign))
summary(aov(mass_day10~rando_assign,data=day5))
day5$tape<-as.character(day5$tape)
##assign names cohort day, tape color, ind_id
day5$uniqueID<-paste("h",day5$cohort_day,substr(day5$tape,1,1),day5$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day5$wellID<-ifelse(day5$rando_assign=="RT eclosion",sample(wells,length(day5$Ind_ID),replace=FALSE),ifelse(day5$rando_assign=="Simulated Overwintering",sample(wells,length(day5$Ind_ID),replace=FALSE),""))
day5$wellID

#write.csv(day5,"../Data/randomizing/2017-10-10_day5_haw_cohort_randomized.csv")

```


### 2017-09-27 cohort day 6
```{r day6}
day6<-read.csv("../Data/randomizing/2017-10-10_day6_haw_cohort_randomization.csv")
day6
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day6$rando_assign<-sample(v,length(day6[,1]),replace=TRUE)


ggplot(day6,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day6$rando_assign))
summary(aov(mass_day10~rando_assign,data=day6))
day6$tape<-as.character(day6$tape)
##assign names cohort day, tape color, ind_id
day6$uniqueID<-paste("h",day6$cohort_day,substr(day6$tape,1,1),day6$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day6$wellID<-ifelse(day6$rando_assign=="RT eclosion",sample(wells,length(day6$Ind_ID),replace=FALSE),ifelse(day6$rando_assign=="Simulated Overwintering",sample(wells,length(day6$Ind_ID),replace=FALSE),""))
day6$wellID

#write.csv(day6,"../Data/randomizing/2017-10-10_day6_haw_cohort_randomized.csv")

```

### 2017-09-28 cohort day 7  

```{r day7}
day7<-read.csv("../Data/randomizing/2017-10-10_day7_haw_cohort_randomization.csv")
day7
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day7$rando_assign<-sample(v,length(day7[,1]),replace=TRUE)


ggplot(day7,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day7$rando_assign))
summary(aov(mass_day10~rando_assign,data=day7))
day7$tape<-as.character(day7$tape)
##assign names cohort day, tape color, ind_id
day7$uniqueID<-paste("h",day7$cohort_day,substr(day7$tape,1,1),day7$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day7$wellID<-ifelse(day7$rando_assign=="RT eclosion",sample(wells,length(day7$Ind_ID),replace=FALSE),ifelse(day7$rando_assign=="Simulated Overwintering",sample(wells,length(day7$Ind_ID),replace=FALSE),""))
day7$wellID

#write.csv(day7,"../Data/randomizing/2017-10-10_day7_haw_cohort_randomized.csv")

```


### 2017-09-29 cohort day 8  

```{r}

day8<-read.csv("../Data/randomizing/2017-10-14_day8_haw_cohort_randomization.csv")
day8
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day8$rando_assign<-sample(v,length(day8[,1]),replace=TRUE)


ggplot(day8,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day8$rando_assign))
summary(aov(mass_day10~rando_assign,data=day8))
day8$tape<-as.character(day8$tape)
##assign names cohort day, tape color, ind_id
day8$uniqueID<-paste("h",day8$cohort_day,substr(day8$tape,1,1),day8$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day8$wellID<-ifelse(day8$rando_assign=="RT eclosion",sample(wells,length(day8$Ind_ID),replace=FALSE),ifelse(day8$rando_assign=="Simulated Overwintering",sample(wells,length(day8$Ind_ID),replace=FALSE),""))
day8$wellID

#write.csv(day8,"../Data/randomizing/2017-10-14_day8_haw_cohort_randomized.csv")
```


### 2017-09-30 cohort day 9  

```{r}

day9<-read.csv("../Data/randomizing/2017-10-14_day9_haw_cohort_randomization.csv")
day9
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day9$rando_assign<-sample(v,length(day9[,1]),replace=TRUE)


ggplot(day9,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day9$rando_assign))
summary(aov(mass_day10~rando_assign,data=day9))
day9$tape<-as.character(day9$tape)
##assign names cohort day, tape color, ind_id
day9$uniqueID<-paste("h",day9$cohort_day,substr(day9$tape,1,1),day9$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day9$wellID<-ifelse(day9$rando_assign=="RT eclosion",sample(wells,length(day9$Ind_ID),replace=FALSE),ifelse(day9$rando_assign=="Simulated Overwintering",sample(wells,length(day9$Ind_ID),replace=FALSE),""))
day9$wellID

#write.csv(day9,"../Data/randomizing/2017-10-14_day9_haw_cohort_randomized.csv")
```


### 2017-10-01 cohort day 10  

```{r}
day10<-read.csv("../Data/randomizing/2017-10-14_day10_haw_cohort_randomization.csv")
day10
v<-c("genetic control","RT eclosion","Simulated Overwintering")
#day10$rando_assign<-sample(v,length(day10[,1]),replace=TRUE)


ggplot(day10,aes(y=mass_day10,x=rando_assign))+geom_boxplot()

summary(as.factor(day10$rando_assign))
summary(aov(mass_day10~rando_assign,data=day10))
day10$tape<-as.character(day10$tape)
##assign names cohort day, tape color, ind_id
day10$uniqueID<-paste("h",day10$cohort_day,substr(day10$tape,1,1),day10$Ind_ID,sep="")



##assign well id for Rt and fridge treatments 
day10$wellID<-ifelse(day10$rando_assign=="RT eclosion",sample(wells,length(day10$Ind_ID),replace=FALSE),ifelse(day10$rando_assign=="Simulated Overwintering",sample(wells,length(day10$Ind_ID),replace=FALSE),""))
day10$wellID

#write.csv(day10,"../Data/randomizing/2017-10-14_day10_haw_cohort_randomized.csv")
```




```{r}
t<-read.csv("../Data/2017-09-29_dataslice.csv")
t$uniqueID<-paste(t$cohort_day,substr(t$tape,1,1),t$Ind_ID,sep="")

y<-read.csv("../Data/2017-10-18_haw_data-slice.csv")
y$uniqueID<-paste("h",y$cohort_day,substr(y$tape,1,1),y$Ind_ID,sep="")
head(y)
write.csv(y,"y.csv")
```
