==> Scripts_analyses/01_script_2018_collection_year_randomization.Rmd <==
---
title: "2018 collection year randomization"
output: html_document
editor_options: 
  chunk_output_type: console
---

This year is a little different because I'm not doing respirometry on all of the samples. Therefore, the uniqueIDs will be a little different. 

For genetic controls, they'll be labeled by host (H or A), cohort day (1-20), then by individual number (1-30ish)



Creating wells

```{r}
v<-c("genetic control","RT eclosion","Simulated Overwintering")
v2<-c("RT eclosion","Simulated Overwintering")
letters<-rep(c("A","B","C","D","E","F","G","H"),12)

num<-sort(rep(seq(1,12,1),8))

wells<-paste(letters,num,sep="")
#wells<-as.factor(wells)
```

```{r}

t<-data.frame(sample(wells,50),treatment=c(rep("RT",25),rep("SO",25)))
t<-data.frame(sample(wells,22),treatment=c(rep("RT",11),rep("SO",11)))
write.csv(t,"test.csv")
write.csv(t,"../Data/randomizing/2018_collection_year/2018-10-12_haw_cohortday2_2018-09-22.csv")

write.csv(t,"../Data/randomizing/2018_collection_year/2018-10-12_haw_cohortday7_2018-09-27.csv")

write.csv(t,"../Data/randomizing/2018_collection_year/2018-10-13_haw_cohortday8_2018-09-28.csv")

write.csv(t,"../Data/randomizing/2018_collection_year/2018-10-14_haw_cohortday9_2018-09-29.csv")

write.csv(t,"../Data/randomizing/2018_collection_year/2018-10-15_haw_cohortday10_2018-09-30.csv")

```

Non respo randomizatiosn 

```{r}
#Sampling for cohort day 6 (2018-08-21)
A6<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A6,"../Data/randomizing/2018_collection_year/2018-09-04_apple_day6_cohort_2018-08-21.csv")

#sampling for cohort day 7 (2018-08-22

A7<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A7,"../Data/randomizing/2018_collection_year/2018-09-04_apple_day7_cohort_2018-08-22.csv")

#sampling for cohort day 8 (2018-08-23
A8<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A8,"../Data/randomizing/2018_collection_year/2018-09-07_apple_day8_cohort_2018-08-23.csv")

#sampling for cohort day 9 (2018-08-24)
A9<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A9,"../Data/randomizing/2018_collection_year/2018-09-07_apple_day9_cohort_2018-08-24.csv")

#sampling for cohort day 10 (2018-08-25)
A10<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A10,"../Data/randomizing/2018_collection_year/2018-09-07_apple_day10_cohort_2018-08-25.csv")

####Apple!!!!####


#Now HAW #######################################################
#Now HAW #######################################################

#sampling for cohort day 6 (2018-09-26)
H6<-data.frame(sample(wells,27*2),treatment=c(rep("RT",27),rep("SO",27)))
write.csv(H6,"../Data/randomizing/2018_collection_year/2018-10-11_haw_day6_cohort_2018-09-26.csv")






##########################################
#2018-08-28 cohort  day 13 apple
##########################################
A13<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A13,"../Data/randomizing/2018_collection_year/2018-09-12_apple_day13_cohort_2018-08-28.csv")

##########################################
#2018-08-29 cohort  day 14 apple
##########################################
A14<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A14,"../Data/randomizing/2018_collection_year/2018-09-13_apple_day14_cohort_2018-08-29.csv")


##########################################
#2018-08-30 cohort  day 15 apple
##########################################
A15<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A15,"../Data/randomizing/2018_collection_year/2018-09-13_apple_day15_cohort_2018-08-30.csv")


##########################################
#2018-08-31 cohort  day 16 apple
##########################################
A16<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A16,"../Data/randomizing/2018_collection_year/2018-09-13_apple_day16_cohort_2018-08-31.csv")

##########################################
#2018-09-01 cohort  day 17 apple
##########################################
A17<-data.frame(sample(wells,60),treatment=c(rep("RT",30),rep("SO",30)))
write.csv(A17,"../Data/randomizing/2018_collection_year/2018-09-13_apple_day17_cohort_2018-09-01.csv")


```

Respo randomization 

Need to distinguish between years

```{r}
##########################################
#2018-08-26 cohort day 11
##########################################

c11<-read.csv("../Data/randomizing/2018_collection_year/2018-09-05_day6cohort_apple_2018-08-21_resp.csv")

c11$uniqueID<-paste("h",c11$cohort_day,substr(c11$tape,1,1),c11$Ind_ID,"_18",sep="")


c11$rando_assign<-sample(v,length(c11[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c11$wellID<-ifelse(c11$rando_assign=="RT eclosion",sample(wells,length(c11$Ind_ID),replace=FALSE),ifelse(c11$rando_assign=="Simulated Overwintering",sample(wells,length(c11$Ind_ID),replace=FALSE),""))
c11$wellID

write.csv(c11,"../Data/randomizing/2018_collection_year/2018-09-05_day11cohort_apple_2018-08-21_resp_randomized.csv")

##########################################
#2018-08-27 cohort  day 12 apple
##########################################
c12<-read.csv("../Data/randomizing/2018_collection_year/2018-09-07_apple_d12_2018-08-27cohort_data.csv")

c12$uniqueID<-paste("a",c12$cohort_day,substr(c12$tape,1,1),c12$Ind_ID,"_18",sep="")


c12$rando_assign<-sample(v,length(c12[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c12$wellID<-ifelse(c12$rando_assign=="RT eclosion",sample(wells,length(c12$Ind_ID),replace=FALSE),ifelse(c12$rando_assign=="Simulated Overwintering",sample(wells,length(c12$Ind_ID),replace=FALSE),""))
c12$wellID

write.csv(c12,"../Data/randomizing/2018_collection_year/2018-09-10_day12cohort_apple_2018-08-27_resp_randomized.csv")


##########################################
#2018-09-08 cohort  day 9+24 apple
##########################################
c924<-read.csv("../Data/randomizing/2018_collection_year/2018-09-19_day10_24cohort_apple_2018-09-08_resp.csv")

c924$uniqueID<-paste("a",c924$cohort_day,substr(c924$tape,1,1),c924$Ind_ID,"_18",sep="")


c924$rando_assign<-sample(v,length(c924[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c924$wellID<-ifelse(c924$rando_assign=="RT eclosion",sample(wells,length(c924$Ind_ID),replace=FALSE),ifelse(c924$rando_assign=="Simulated Overwintering",sample(wells,length(c924$Ind_ID),replace=FALSE),""))
c924$wellID

write.csv(c924,"../Data/randomizing/2018_collection_year/2018-09-19_day8_24cohort_apple_2018-09-08_resp_randomized.csv")


##########################################
#2018-09-09 cohort  day 10+25 apple
##########################################
c1025<-read.csv("../Data/randomizing/2018_collection_year/2018-09-19_day10_25cohort_apple_2018-09-09_resp.csv")

c1025$uniqueID<-paste("a",c1025$cohort_day,substr(c1025$tape,1,1),c1025$Ind_ID,"_18",sep="")


c1025$rando_assign<-sample(v,length(c1025[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c1025$wellID<-ifelse(c1025$rando_assign=="RT eclosion",sample(wells,length(c1025$Ind_ID),replace=FALSE),ifelse(c1025$rando_assign=="Simulated Overwintering",sample(wells,length(c1025$Ind_ID),replace=FALSE),""))
c1025$wellID

write.csv(c1025,"../Data/randomizing/2018_collection_year/2018-09-19_day10_25cohort_apple_2018-09-09_resp_randomized.csv")

##########################################
#2018-09-10 cohort  day 11+26 apple
##########################################
c1126<-read.csv("../Data/randomizing/2018_collection_year/2018-09-20_day11_26cohort_apple_2018-09-10_resp.csv")

c1126$uniqueID<-paste("a",c1126$cohort_day,substr(c1126$tape,1,1),c1126$Ind_ID,"_18",sep="")


c1126$rando_assign<-sample(v,length(c1126[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c1126$wellID<-ifelse(c1126$rando_assign=="RT eclosion",sample(wells,length(c1126$Ind_ID),replace=FALSE),ifelse(c1126$rando_assign=="Simulated Overwintering",sample(wells,length(c1126$Ind_ID),replace=FALSE),""))
c1126$wellID

write.csv(c1126,"../Data/randomizing/2018_collection_year/2018-09-20_day11_26cohort_apple_2018-09-10_resp_randomized.csv")

##########################################
#2018-09-11 cohort  day 12+27 apple
##########################################
c1227<-read.csv("../Data/randomizing/2018_collection_year/2018-09-21_day12_27cohort_apple_2018-09-11_resp.csv")

c1227$uniqueID<-paste("a",c1227$cohort_day,substr(c1227$tape,1,1),c1227$Ind_ID,"_18",sep="")


c1227$rando_assign<-sample(v,length(c1227[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c1227$wellID<-ifelse(c1227$rando_assign=="RT eclosion",sample(wells,length(c1227$Ind_ID),replace=FALSE),ifelse(c1227$rando_assign=="Simulated Overwintering",sample(wells,length(c1227$Ind_ID),replace=FALSE),""))
c1227$wellID

write.csv(c1227,"../Data/randomizing/2018_collection_year/2018-09-21_day12_27cohort_apple_2018-09-11_resp_randomized.csv")


##########################################
#2018-09-22 cohort  day 2 HAW
##########################################
c2h<-read.csv("../Data/randomizing/2018_collection_year/2018-10-01_day2_cohort_haw_2018-09-22_resp.csv")

c2h$uniqueID<-paste("a",c2h$cohort_day,substr(c2h$tape,1,1),c2h$Ind_ID,"_18",sep="")


c2h$rando_assign<-sample(v,length(c2h[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c2h$wellID<-ifelse(c2h$rando_assign=="RT eclosion",sample(wells,length(c2h$Ind_ID),replace=FALSE),ifelse(c2h$rando_assign=="Simulated Overwintering",sample(wells,length(c2h$Ind_ID),replace=FALSE),""))
c2h$wellID

write.csv(c2h,"../Data/randomizing/2018_collection_year/2018-10-01_day2_cohort_haw_2018-09-22_resp_randomized.csv")

##########################################
#2018-09-23 cohort  day 3 HAW
##########################################
c3h<-read.csv("../Data/randomizing/2018_collection_year/2018-10-03_day3_cohort_haw_2018-09-23_resp.csv")

c3h$uniqueID<-paste("a",c3h$cohort_day,substr(c3h$tape,1,1),c3h$Ind_ID,"_18",sep="")


c3h$rando_assign<-sample(v,length(c3h[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c3h$wellID<-ifelse(c3h$rando_assign=="RT eclosion",sample(wells,length(c3h$Ind_ID),replace=FALSE),ifelse(c3h$rando_assign=="Simulated Overwintering",sample(wells,length(c3h$Ind_ID),replace=FALSE),""))
c3h$wellID

write.csv(c3h,"../Data/randomizing/2018_collection_year/2018-10-03_day3_cohort_haw_2018-09-23_resp_randomized.csv")

##########################################
#2018-09-24 cohort  day 4 HAW
##########################################
c4h<-read.csv("../Data/randomizing/2018_collection_year/2018-10-05_day4_cohort_haw_2018-09-24_resp.csv")

c4h$uniqueID<-paste("a",c4h$cohort_day,substr(c4h$tape,1,1),c4h$Ind_ID,"_18",sep="")


c4h$rando_assign<-sample(v,length(c4h[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c4h$wellID<-ifelse(c4h$rando_assign=="RT eclosion",sample(wells,length(c4h$Ind_ID),replace=FALSE),ifelse(c4h$rando_assign=="Simulated Overwintering",sample(wells,length(c4h$Ind_ID),replace=FALSE),""))
c4h$wellID

write.csv(c4h,"../Data/randomizing/2018_collection_year/2018-10-05_day4_cohort_haw_2018-09-24_resp_randomized.csv")


##########################################
#2018-09-25 cohort  day 5 HAW
##########################################
c5h<-read.csv("../Data/randomizing/2018_collection_year/2018-10-05_day5_cohort_haw_2018-09-25_resp.csv")

c5h$uniqueID<-paste("a",c5h$cohort_day,substr(c5h$tape,1,1),c5h$Ind_ID,"_18",sep="")


c5h$rando_assign<-sample(v,length(c5h[,1]),replace=TRUE)
##assign well id for Rt and fridge treatments 
c5h$wellID<-ifelse(c5h$rando_assign=="RT eclosion",sample(wells,length(c5h$Ind_ID),replace=FALSE),ifelse(c5h$rando_assign=="Simulated Overwintering",sample(wells,length(c5h$Ind_ID),replace=FALSE),""))
c5h$wellID

write.csv(c5h,"../Data/randomizing/2018_collection_year/2018-10-05_day5_cohort_haw_2018-09-25_resp_randomized.csv")
```

```{r}
write.csv(rbind(cbind(sample(v2,30,replace = TRUE),sample(wells,30,replace=FALSE)),cbind(sample(v2,30,replace = TRUE),sample(wells,30,replace=FALSE))),"test.csv")

```
