==> Scripts_analyses/02_01_script_2018-02-27_eclosion_data_explore.Rmd <==
---
title: "2018-02-27_eclosion_summaries"
output:
  pdf_document: 
    highlight: monochrome
    toc: yes
  html_document: default
editor_options: 
  chunk_output_type: console
---

# Eclosion data 

```{r}
library(data.table)
library(ggplot2)
library(dplyr)
library(lubridate)

#ggplot graph settings
T<-theme_bw()+theme(,text=element_text(size=25),axis.text=element_text(size=25), panel.grid.major=element_blank(), panel.grid.minor.x = element_blank(), panel.grid = element_blank(), legend.key = element_blank(),axis.title.y=element_text(margin=margin(t=0,r=15,b=0,l=0)),axis.title.x=element_text(margin=margin(t=15,r=,b=0,l=0)))+ theme(legend.position="none")
```

## Loading in the data

```{r}
##grabbing data if git cloned
#eclosions<-fread("../Data/2018-01-26_rhagoletis_masterdata_data_slice.csv")
#grabbing from online
eclosions<-fread("../Data/2018-08-14_master_data_slice.csv")

head(eclosions,10)
names(eclosions)
dim(eclosions)
#tail(eclosions,10)


eclosions%>%
  group_by(Host,Site_name)%>%
  summarise(sample_size=length(Host))%>%
  filter(Site_name!="Blank")%>%
  knitr::kable()

### getting numbers for eclosions without simulated overwintering, so these pupae were left at rearing temperature
t1<-eclosions%>%
  dplyr::filter(eclosion_days>1)%>%
  dplyr::group_by(Host,organism,treatment)%>%
  dplyr::summarise(counts=length(eclosion_days))
t1

knitr::kable(t1)

#2018-04-26 notes
#there is a apple host with no orgnaism
#eclosions%>%
 # filter(eclosion_days>1 & organism =="")%>%
  #select(uniqueID)
#eclosed on 2017-11-30; nb 002 pg 19; designating as a fly; it eclosed but died same day

#Sample missing treatment 
#eclosions%>%
 # filter(eclosion_days>1 & Host == "Apple"&organism=="fly",treatment=="")%>%
  #select(uniqueID)
  
  
```

## plotting distributions with Ggplot

```{r}
#histogram
treat.filt<-eclosions%>%
  filter(treatment!="" & treatment !="GC" & treatment !="F1"&organism!="")

#ggplot(treat.filt,aes(x=eclosion_days,fill=paste(Host,organism)))+geom_histogram(position="identity",alpha=.52)
ggplot(treat.filt,aes(x=eclosion_days,fill=paste(Host,organism)))+geom_density(position="identity",alpha=.52)+facet_grid(treatment~organism)

ggplot(treat.filt,aes(x=eclosion_days,fill=paste(Host)))+geom_histogram(position="identity",alpha=.52)+facet_grid(treatment~organism)+xlab("Eclosion(days)")


#density of flies only
#flies<-subset(eclosions,organism=="fly")
flies<-eclosions%>%
  filter(treatment!="" & treatment !="GC" & treatment !="F1"&organism=="fly")

ggplot(flies,aes(x=eclosion_days,fill=paste(Host)))+geom_density(position="identity",alpha=.52)+facet_grid(treatment~.)

ggplot(flies,aes(x=eclosion_days,fill=paste(Host)))+geom_histogram(position="identity",alpha=.52)+facet_grid(treatment~.)

##instead of plotting distributions, let's plot averages

ggplot(treat.filt,aes(y=eclosion_days,x=Host))+geom_boxplot()+facet_grid(treatment~organism,scales="free")+ylab("Eclosion (days)")
#lets try some stats

#on the flies only dataset
mod1<-aov(eclosion_days~Host*treatment,data=flies)
summary(mod1)


##simulated overwintering
ap.ks<-subset(flies,Host=="Apple"&treatment=="SO")
haw.ks<-subset(flies,Host=="Haw"&treatment=="SO")
ks.test(ap.ks$eclosion_days,haw.ks$eclosion_days)

###Rearing temperature
rtap.ks<-subset(flies,Host=="Apple"&treatment=="RT")
rthaw.ks<-subset(flies,Host=="Haw"&treatment=="RT")
ks.test(rtap.ks$eclosion_days,rthaw.ks$eclosion_days)

```

# Skipped eclosion data

1. Load in the data from skipped cohorts 
  * Use this link: https://raw.githubusercontent.com/adnguyen/Circadian_rhythm_runs_seasonal_timing/master/Data/2017-10-22_skipped_apple_cohort_eclosions.csv
  
2. Plot (using ggplot) a histogram of eclosion_days 

```{r}

skipped<-fread("https://raw.githubusercontent.com/adnguyen/Circadian_rhythm_runs_seasonal_timing/master/Data/2017-10-22_skipped_apple_cohort_eclosions.csv")


str(skipped)

ggplot(skipped,aes(x=Eclosion_day,fill=Host))+geom_histogram(position="identity",alpha=.5)
ggplot(skipped,aes(x=Eclosion_day,fill=Host))+geom_density(position="identity",alpha=.5)
```



# Quantifying days until eclosion

```{r}
glimpse(eclosions)
#mdy(eclosions$exit_fridge_date)

eclosions$Eclosion_reference_date
eclosions$eclosion_date

##subtracting date of eclosion from day 15 treatment for RT and out of winter date for SO treatment to estimate the number of days until eclosions
#BOTH RT and SO don't have the 15 day pre-winter time, so I'm adding it on to the end
eclosions$new.eclosions<-difftime(as.Date(eclosions$eclosion_date),as.Date(eclosions$Eclosion_reference_date),units=c("days"))+15

#ggplot(eclosions,aes(x=new.eclosions,eclosion_days))+geom_point()

```

# Comparing flies and wasps


```{r}
ecl.filter<-eclosions%>%
  filter(treatment!=""&treatment!="GC",organism!="")

ggplot(ecl.filter,aes(x=new.eclosions,fill=paste(Host,organism)))+geom_density(position="identity",alpha=.52)+facet_grid(treatment~organism)

ggplot(ecl.filter,aes(x=new.eclosions,fill=paste(Host)))+geom_histogram(position="identity",alpha=.52)+facet_grid(treatment~organism)+xlab("Eclosion(days)")

ggplot(ecl.filter,aes(x=new.eclosions,fill=paste(Host,organism)))+geom_density(position="identity",alpha=.52)+facet_grid(treatment~.)+xlab("Eclosion(days)")


## overall sample sizes

ecl.filter%>%
  group_by(Host,organism,treatment)%>%
  summarise(n=length(Host))

###sample sizes for RT
RT.sub<-subset(ecl.filter,treatment=="RT")
sample.sizes<-RT.sub%>%
  group_by(Host,organism)%>%
  summarise(n=length(Host))
#ggplot(sample.sizes,aes(y=n,x=paste(Host,organism)))+geom_bar()

# samples for RT and haw
RT.sub.haw<-subset(ecl.filter,treatment=="RT"& Host=="Haw")
RT.sub.haw%>%
  group_by(Host,organism)%>%
  summarise(n=length(Host))


mod6<-aov(new.eclosions~organism,data=RT.sub.haw)
summary(mod6)

#split out plot by host and plot organism within each host
ggplot(ecl.filter,aes(x=new.eclosions,fill=paste(Host,organism)))+geom_density(position="identity",alpha=.52)+facet_grid(treatment~Host)+xlab("Eclosion(days)")+T

##boxplots
ggplot(ecl.filter,aes(y=eclosion_days,x=Host,fill=organism))+geom_boxplot()+facet_grid(treatment~.,scales="free")+ylab("Adult Emergence (days)")+T
##violin plot
ggplot(ecl.filter,aes(y=eclosion_days,x=Host,fill=organism))+geom_violin()+facet_grid(treatment~.,scales="free")+ylab("Adult Emergence (days)")+T



#what are the sample sizes
ecl.filter%>%
  group_by(treatment,organism,Host)%>%
  summarise(n=length(Host))


##for the biological rhythms conference, I just need to take the RT for Haw samples
ecl.filter.haw<-ecl.filter%>%
  filter(Host=="Haw"&treatment=="RT")
ggplot(ecl.filter.haw,aes(y=eclosion_days,x=organism,fill=organism))+geom_boxplot()+T+ylab("Adult Emergence (days)")
```


# Just analyzing flies 

```{r}
subecl<-eclosions%>%
  filter(organism=="fly"&treatment!="" &treatment !="GC")

mod2<-aov(new.eclosions~Host*treatment,data=subecl)
summary(mod2)

#ggplot(subecl,aes(y=new.eclosions,x=Host))+geom_boxplot()+facet_grid(.~treatment)

ggplot(subecl,aes(x=eclosion_days,fill=paste(Host)))+geom_density(position="identity",alpha=.52)+facet_grid(treatment~.)

ggplot(subecl,aes(x=eclosion_days,fill=paste(Host)))+geom_histogram(position="identity",alpha=.52)+facet_grid(treatment~.)

###comparing just the SO 
subecl2<-subecl%>%
  filter(treatment=="SO")
mod22<-aov(new.eclosions~Host,data=subecl2)
summary(mod22)
ggplot(subecl2,aes(x=Host,y=new.eclosions))+geom_boxplot()

subecl2%>%
  group_by(Host)%>%
  summarise(n=length(Host))

#comparing similar sites only 
subecl.site<-subecl%>%
  filter(Site_name!="Milford"& Site_name!="Coll Hall" & Site_name!="MSU")

mod3<-aov(new.eclosions~Host*treatment,data=subecl.site)
summary(mod3)
ggplot(subecl.site,aes(x=eclosion_days,fill=paste(Host)))+geom_density(position="identity",alpha=.52)+facet_grid(treatment~Site_name)

```

### Write out processed dataset

```{r}
write.csv(eclosions,"05_master_dataset_eclosions_IDs.csv")
```





# 2018-2019 dataset 

```{r 2019 dataset}
dat8<-fread("../Data/data-slice_2019-01-08_master_datac_2018_collection_year.csv")

glimpse(dat8)
#unique(levels(as.factor(dat8$collection_date)))

dat8$treatment15_5months<-ymd(dat8$treatment_day15)%m+%months(5)

five.month.schedule<-dat8%>%
  group_by(collection_date,cohort_date,Host,treatment_day15,treatment15_5months)%>%
  filter(Host!="Crab Apple" & Host!="Crab")%>%
  summarize(months5=unique(treatment15_5months))%>%
  na.omit()
  
fwrite(five.month.schedule,"2018-2019_collection_year_SO_schedule.csv")


dat8%>%
  filter(cohort_date=="2018-12-02" & Host=="Apple")#%>%
  #write.csv("blanks.csv")
```
