==> Scripts_analyses/03_script_2017-11-19_processing_trik_monitor_data.Rmd <==
---
title: "Processing monitor data"
author: "Andrew D. Nguyen"
date: "11/19/2017"
output: 
  html_document: 
    toc: yes
editor_options: 
  chunk_output_type: console
---

# Processing monitor data

Take raw counts which is in different monitors and conditions and combine them. 


## Load libraries
```{r loading libraries}
library(ggplot2)
library(dplyr)
library(lubridate)
library(data.table)
library(tidyr)
library(chron)
```

## Load in monitor data   

* add header names to the data   
* for every unique ID, extract behavioral counts

2018-04-18
**additional notes:** 

We were running low on trikinetics space, so we had to switch some monitors around. First, #1 and 2 are in entrainment, while 3,4,5,6 are in free-run conditions. We took out #6 from free run to entrainment on 2018-03-23,20:00 or 8:00PM (nb#002,page 178, parse on 227487) and then switched it back on 2018-04-05(parse on count number 241326; nb#003,page14-15). 

To deal with monitor #6 having a combination of free-run and entrainment data counts, we will need to subset out the data from when #6 was in entrainment. We can take this subset and call it a diff monitor altogether ("6.2") and then assign different experimental classes for #6 and #6.2. 

Another complication: There are samples in the move from entrainment to free-run in monitor #6 that needed to be moved to monitors #1 and #2. So this means some flies have had entrainment that were in 2 monitors. See notebook #003, page 14-15. Here is the list: 

```{r table of samples with 2 entrainment trik monitores,echo=FALSE}
knitr::kable(fread("../Data/raw/Trikinetics/2018-04-18_nb3_pg14_multiple_entrainments_trik.csv"))


```




### Trikinetics monitor data processing

```{r load monitor data}
######################################################
######################################################
#m1<-read.table("../Data/raw/Trikinetics/Monitor1.txt")
m1<-read.table("../Data/raw/Trikinetics/Monitor1_2018-07-31.txt")
m1$monitor<-rep(1,length(m1$V1))
#m2<-read.table("../Data/raw/Trikinetics/Monitor2.txt")
m2<-read.table("../Data/raw/Trikinetics/Monitor2_2018_07-31.txt")
m2$monitor<-rep(2,length(m2$V1))
m3<-read.table("../Data/raw/Trikinetics/Monitor3.txt")
m3$monitor<-rep(3,length(m3$V1))
m4<-read.table("../Data/raw/Trikinetics/Monitor4.txt")
m4$monitor<-rep(4,length(m4$V1))
m5<-read.table("../Data/raw/Trikinetics/Monitor5.txt")
m5$monitor<-rep(5,length(m5$V1))
#monitor 6 needs to be handled differently because there is a mixture of free-run and entrainment data

m6<-read.table("../Data/raw/Trikinetics/Monitor6.txt")

#grabbing entrainment
m6.1<-m6%>%
  filter(V1<227487 | V1 > 241326)
m6.1$monitor<-rep(6,length(m6.1$V1))
#grabbing free-run
m6.2<-m6%>%
  filter(V1>227487 & V1 < 241326)
m6.2$monitor<-rep(7,length(m6.2$V1)) ### setting free run data for this monitor as number 7
######################################################
######################################################

#allm<-rbind(m1,m2,m3,m4,m5,m6) # combining all of the monitor data
allm<-rbind(m1,m2,m3,m4,m5,m6.1,m6.2) # combining all of the monitor data
allm$V3<-match(allm$V3,month.abb) # converint month to numeric


allm$V2<-paste("20",allm$V4,"-",allm$V3,"-",formatC(allm$V2,width=2,flag="0"),sep="") # reformatting the date

#parse out values that are not 1 (not reading) in V6 ( it is column 4 in the trik manaul but the dates are separated when reading into R)
allm<-allm%>%
  filter(V6==1)

allm<-allm[,-3:-4]# excluding time we dont need
allm<-allm[,-5:-9]# excluding time we dont need
head(allm)


#names(allm)<-c("time_step","date","time","trik_status","none1","none2","none3","none4","none5","light_status",paste("t",seq(1,32),sep=""),"monitor")

names(allm)<-c("time_step","date","time","trik_status","light_status",paste("t",seq(1,32),sep=""),"monitor")
names(allm)

allm$fulltime<-ymd_hms(paste(allm$date,allm$time,sep=""),tz="US/Eastern")

#t1 - t32 = 6-37 
#writing out the dataset
fwrite(allm,"../Data/raw/Trikinetics/01_2018-04-18_allm_all_monitorscombined.csv")###wide format; gitignore because file is too large (> 100mb)
######################################################
######################################################
```

# Manipulating monitor data  

converting from wide to long format

```{r manipulating monitor data}
#################################################

monitors.long<-gather(allm,position,counts,t1:t32)
monitors.long$position<-as.numeric(substr(monitors.long$position,2,4))
head(monitors.long)
#monitors.long$fulltime<-ymd_hms(head(monitors.long$fulltime),tz="US/Eastern")
monitors.long$fulltime<-chron(dates=monitors.long$date,times=monitors.long$time,format=c('y-m-d','h:m:s'))
# write out dataset, long format
fwrite(monitors.long,"../Data/raw/Trikinetics/02_2018-04-18_trik_dat_long.csv")
#################################################

```

# testing out whether we can read data from URL

```{r eval=FALSE}

test<-fread("https://raw.githubusercontent.com/adnguyen/Circadian_rhythm_runs_seasonal_timing/master/Data/2017-08-24_rhagoletis_data_sheet_2017-12-13_dataslice.csv")
head(test)
#entrainment time 1
#test$et1<-ymd_hm(paste(test$eclosion_date,test$Trikinetics_entry_LD_time,sep=" "),tz="US/Eastern")
test$et1<-chron(dates=test$eclosion_date,times=test$Trikinetics_entry_LD_time,format=c('y-m-d','h:m:s'))

#entrainment time 2
#test$et2<-ymd_hm(paste(test$Trikinetic_exit_date,test$Trikinetics_exit_LD_time,sep=" "),tz="US/Eastern")
test$et2<-chron(dates=test$Trikinetic_exit_date,times=test$Trikinetics_exit_LD_time,format=c('y-m-d','h:m:s'))

test$int1<-interval(test$et1,test$et2)
str(test)



```

# Manually parsing out a subset of individuals

**h4o4**  

```{r h4o4}
#############################################
###entrainment
#############################################

### grabbing the data from allm dataset to get the time step, date, time, trik status, light status, monitor, and monitor position for h4o4

###we need to look up and specify the time step that the animal was in the trikinetics rig
h4o4<-subset(allm,monitor =="1" & time_step <29241 & time_step>20964, select=c("time_step","date","time","trik_status","light_status","monitor","t1")) 
#dim(h4o4)

#specifying the entrainment part of data
h4o4$experiment<-rep("Entrainment",length(h4o4$time_step))

##recoding the trik position (needs to be specific to the animal )
h4o4$trik_position<-rep(1,length(h4o4$time_step))

## renaming trik position to counts
names(h4o4)[7]<-c("counts") 


#############################################
### free run
#############################################
h4o4f<-subset(allm,monitor =="4" & time_step <52917 & time_step>29242, select=c("time_step","date","time","trik_status","light_status","monitor","t19"))
dim(h4o4f)
h4o4f$experiment<-rep("Free-run",length(h4o4f$time_step))
h4o4f$trik_position<-rep(19,length(h4o4f$time_step))
names(h4o4f)[7]<-c("counts")


#### stitching the data together  
h4o4full<-rbind(h4o4,h4o4f)
h4o4full$uniqueID<-rep("h4o4",length(h4o4full$time_step))
head(h4o4full)
```



**h2b25**

```{r h2b25}
###entrainment
h2b25<-subset(allm,monitor =="1" & time_step <30650 & time_step>22063, select=c("time_step","date","time","trik_status","light_status","monitor","t14")) 

h2b25$experiment<-rep("Entrainment",length(h2b25$time_step))

h2b25$trik_position<-rep(14,length(h2b25$time_step))

names(h2b25)[7]<-c("counts") 

###free run
h2b25f<-subset(allm,monitor =="4" & time_step <52917 & time_step>30650, select=c("time_step","date","time","trik_status","light_status","monitor","t23"))
dim(h2b25f)
h2b25f$experiment<-rep("Free-run",length(h2b25f$time_step))
h2b25f$trik_position<-rep(23,length(h2b25f$time_step))
names(h2b25f)[7]<-c("counts")

###stitching
h2b25full<-rbind(h2b25,h2b25f)
h2b25full$uniqueID<-rep("h2b25",length(h2b25full$time_step))
head(h2b25full)
```

**h5o22**
```{r h5o22}
###entrainment
h5o22<-subset(allm,monitor =="2" & time_step <34262 & time_step>26051, select=c("time_step","date","time","trik_status","light_status","monitor","t22")) 

h5o22$experiment<-rep("Entrainment",length(h5o22$time_step))

h5o22$trik_position<-rep(22,length(h5o22$time_step))

names(h5o22)[7]<-c("counts") 

###free run
h5o22f<-subset(allm,monitor =="3" & time_step <52917 & time_step>34264, select=c("time_step","date","time","trik_status","light_status","monitor","t26"))
dim(h5o22f)
h5o22f$experiment<-rep("Free-run",length(h5o22f$time_step))
h5o22f$trik_position<-rep(26,length(h5o22f$time_step))
names(h5o22f)[7]<-c("counts")

###stitching
h5o22full<-rbind(h5o22,h5o22f)
h5o22full$uniqueID<-rep("h5o22",length(h5o22full$time_step))
head(h5o22full)
```

**12b6**
```{r 12b6}
###entrainment
a12b6<-subset(allm,monitor =="1" & time_step <22157 & time_step>15806, select=c("time_step","date","time","trik_status","light_status","monitor","t17")) 

a12b6$experiment<-rep("Entrainment",length(a12b6$time_step))

a12b6$trik_position<-rep(17,length(a12b6$time_step))

names(a12b6)[7]<-c("counts") 

###free run
a12b6f<-subset(allm,monitor =="4" & time_step <52917 & time_step>22157, select=c("time_step","date","time","trik_status","light_status","monitor","t19"))
dim(h2b25f)
a12b6f$experiment<-rep("Free-run",length(a12b6f$time_step))
a12b6f$trik_position<-rep(19,length(a12b6f$time_step))
names(a12b6f)[7]<-c("counts")

###stitching
a12b6full<-rbind(a12b6,a12b6f)
a12b6full$uniqueID<-rep("a12b6",length(a12b6full$time_step))
head(a12b6full)
```

**5w73**
```{r 5w73}
a5w73<-subset(allm,monitor =="2" & time_step <34905 & time_step>27492, select=c("time_step","date","time","trik_status","light_status","monitor","t11")) 

a5w73$experiment<-rep("Entrainment",length(a5w73$time_step))

a5w73$trik_position<-rep(11,length(a5w73$time_step))

names(a5w73)[7]<-c("counts") 

###free run
a5w73f<-subset(allm,monitor =="5" & time_step <52917 & time_step>34905, select=c("time_step","date","time","trik_status","light_status","monitor","t22"))
dim(a5w73f)
a5w73f$experiment<-rep("Free-run",length(a5w73f$time_step))
a5w73f$trik_position<-rep(22,length(a5w73f$time_step))
names(a5w73f)[7]<-c("counts")

###stitching
a5w73full<-rbind(a5w73,a5w73f)
a5w73full$uniqueID<-rep("a5w73",length(a5w73full$time_step))
head(a5w73full)
```

**4w66**
```{r 4w66}
a4w66<-subset(allm,monitor =="2" & time_step <18896 & time_step>893, select=c("time_step","date","time","trik_status","light_status","monitor","t23")) 

a4w66$experiment<-rep("Entrainment",length(a4w66$time_step))

a4w66$trik_position<-rep(23,length(a4w66$time_step))

names(a4w66)[7]<-c("counts") 

###free run
a4w66f<-subset(allm,monitor =="3" & time_step <33980 & time_step>18896, select=c("time_step","date","time","trik_status","light_status","monitor","t4"))
dim(a4w66f)
a4w66f$experiment<-rep("Free-run",length(a4w66f$time_step))
a4w66f$trik_position<-rep(4,length(a4w66f$time_step))
names(a4w66f)[7]<-c("counts")

###stitching
a4w66full<-rbind(a4w66,a4w66f)
a4w66full$uniqueID<-rep("a4w66",length(a4w66full$time_step))
head(a4w66full)
```

**10w12**
```{r 10w12}
a10w12<-subset(allm,monitor =="2" & time_step <18896 & time_step>8120, select=c("time_step","date","time","trik_status","light_status","monitor","t10")) 

a10w12$experiment<-rep("Entrainment",length(a10w12$time_step))

a10w12$trik_position<-rep(10,length(a10w12$time_step))

names(a10w12)[7]<-c("counts") 

###free run
a10w12f<-subset(allm,monitor =="3" & time_step <52917 & time_step>18896, select=c("time_step","date","time","trik_status","light_status","monitor","t27"))
dim(a10w12f)
a10w12f$experiment<-rep("Free-run",length(a10w12f$time_step))
a10w12f$trik_position<-rep(27,length(a10w12f$time_step))
names(a10w12f)[7]<-c("counts")

###stitching
a10w12full<-rbind(a10w12,a10w12f)
a10w12full$uniqueID<-rep("a10w12",length(a10w12full$time_step))
head(a10w12full)
```

**10o49**
```{r 10o49}
a10o49<-subset(allm,monitor =="2" & time_step <18896 & time_step>8119, select=c("time_step","date","time","trik_status","light_status","monitor","t2")) 

a10o49$experiment<-rep("Entrainment",length(a10o49$time_step))

a10o49$trik_position<-rep(2,length(a10o49$time_step))

names(a10o49)[7]<-c("counts") 

###free run
a10o49f<-subset(allm,monitor =="3" & time_step <52917 & time_step>18896, select=c("time_step","date","time","trik_status","light_status","monitor","t28"))
dim(a10o49f)
a10o49f$experiment<-rep("Free-run",length(a10o49f$time_step))
a10o49f$trik_position<-rep(28,length(a10o49f$time_step))
names(a10o49f)[7]<-c("counts")

###stitching
a10o49full<-rbind(a10o49,a10o49f)
a10o49full$uniqueID<-rep("a10o49",length(a10o49full$time_step))
head(a10o49full)
```

**12w40**
```{r 12w40}
a12w40<-subset(allm,monitor =="1" & time_step <19477 & time_step>10937, select=c("time_step","date","time","trik_status","light_status","monitor","t8")) 

a12w40$experiment<-rep("Entrainment",length(a12w40$time_step))

a12w40$trik_position<-rep(8,length(a12w40$time_step))

names(a12w40)[7]<-c("counts") 

###free run
a12w40f<-subset(allm,monitor =="3" & time_step <52917 & time_step>19460, select=c("time_step","date","time","trik_status","light_status","monitor","t30"))
dim(a12w40f)
a12w40f$experiment<-rep("Free-run",length(a12w40f$time_step))
a12w40f$trik_position<-rep(30,length(a12w40f$time_step))
names(a12w40f)[7]<-c("counts")

###stitching
a12w40full<-rbind(a12w40,a12w40f)
a12w40full$uniqueID<-rep("a12w40",length(a12w40full$time_step))
head(a12w40full)
```

**12b43**
```{r 12b43}
a12b43<-subset(allm,monitor =="1" & time_step <19477 & time_step>10937, select=c("time_step","date","time","trik_status","light_status","monitor","t15")) 

a12b43$experiment<-rep("Entrainment",length(a12b43$time_step))

a12b43$trik_position<-rep(15,length(a12b43$time_step))

names(a12b43)[7]<-c("counts") 

###free run
a12b43f<-subset(allm,monitor =="3" & time_step <52917 & time_step>19460, select=c("time_step","date","time","trik_status","light_status","monitor","t29"))
dim(a12b43f)
a12b43f$experiment<-rep("Free-run",length(a12b43f$time_step))
a12b43f$trik_position<-rep(29,length(a12b43f$time_step))
names(a12b43f)[7]<-c("counts")

###stitching
a12b43full<-rbind(a12b43,a12b43f)
a12b43full$uniqueID<-rep("a12b43",length(a12b43full$time_step))
head(a12b43full)
```

**13o28**
```{r 13o28}
a13o28<-subset(allm,monitor =="1" & time_step <19829 & time_step>12424, select=c("time_step","date","time","trik_status","light_status","monitor","t7")) 

a13o28$experiment<-rep("Entrainment",length(a13o28$time_step))

a13o28$trik_position<-rep(7,length(a13o28$time_step))

names(a13o28)[7]<-c("counts") 

###free run
a13o28f<-subset(allm,monitor =="4" & time_step <34629 & time_step>19829, select=c("time_step","date","time","trik_status","light_status","monitor","t6"))
dim(a13o28f)
a13o28f$experiment<-rep("Free-run",length(a13o28f$time_step))
a13o28f$trik_position<-rep(6,length(a13o28f$time_step))
names(a13o28f)[7]<-c("counts")

###stitching
a13o28full<-rbind(a13o28,a13o28f)
a13o28full$uniqueID<-rep("a13o28",length(a13o28full$time_step))
head(a13o28full)
```

**13o11**
```{r 13o11}
a13o11<-subset(allm,monitor =="2" & time_step <26421 & time_step>18899, select=c("time_step","date","time","trik_status","light_status","monitor","t1")) 

a13o11$experiment<-rep("Entrainment",length(a13o11$time_step))

a13o11$trik_position<-rep(1,length(a13o11$time_step))

names(a13o11)[7]<-c("counts") 

###free run
a13o11f<-subset(allm,monitor =="4" & time_step <52917 & time_step>26406, select=c("time_step","date","time","trik_status","light_status","monitor","t24"))
dim(a13o11f)
a13o11f$experiment<-rep("Free-run",length(a13o11f$time_step))
a13o11f$trik_position<-rep(24,length(a13o11f$time_step))
names(a13o11f)[7]<-c("counts")

###stitching
a13o11full<-rbind(a13o11,a13o11f)
a13o11full$uniqueID<-rep("a13o11",length(a13o11full$time_step))
head(a13o11full)
```

**5b25**
```{r 5b25}
a5b25<-subset(allm,monitor =="1" & time_step <34638 & time_step>26812, select=c("time_step","date","time","trik_status","light_status","monitor","t26"))  

a5b25$experiment<-rep("Entrainment",length(a5b25$time_step))

a5b25$trik_position<-rep(26,length(a5b25$time_step))

names(a5b25)[7]<-c("counts") 

###free run
a5b25f<-subset(allm,monitor =="5" & time_step <52917 & time_step>34349, select=c("time_step","date","time","trik_status","light_status","monitor","t32"))
dim(a5b25f)
a5b25f$experiment<-rep("Free-run",length(a5b25f$time_step))
a5b25f$trik_position<-rep(32,length(a5b25f$time_step))
names(a5b25f)[7]<-c("counts")

###stitching
a5b25full<-rbind(a5b25,a5b25f)
a5b25full$uniqueID<-rep("a5b25",length(a5b25full$time_step))
head(a5b25full)
```

**10o73**
```{r 10o73}
a10o73<-subset(allm,monitor =="1" & time_step <37780 & time_step>30039, select=c("time_step","date","time","trik_status","light_status","monitor","t3"))  

a10o73$experiment<-rep("Entrainment",length(a10o73$time_step))

a10o73$trik_position<-rep(3,length(a10o73$time_step))

names(a10o73)[7]<-c("counts") 

###free run
a10o73f<-subset(allm,monitor =="4" & time_step <52917 & time_step>37804, select=c("time_step","date","time","trik_status","light_status","monitor","t1"))
dim(a10o73f)
a10o73f$experiment<-rep("Free-run",length(a10o73f$time_step))
a10o73f$trik_position<-rep(1,length(a10o73f$time_step))
names(a10o73f)[7]<-c("counts")

###stitching
a10o73full<-rbind(a10o73,a10o73f)
a10o73full$uniqueID<-rep("a10o73",length(a10o73full$time_step))
head(a10o73full)
```

**3o51**
```{r 3o51}
a3o51<-subset(allm,monitor =="1" & time_step <34349 & time_step>26812, select=c("time_step","date","time","trik_status","light_status","monitor","t17"))  

a3o51$experiment<-rep("Entrainment",length(a3o51$time_step))

a3o51$trik_position<-rep(17,length(a3o51$time_step))

names(a3o51)[7]<-c("counts") 

###free run
a3o51f<-subset(allm,monitor =="5" & time_step <52917 & time_step>34349, select=c("time_step","date","time","trik_status","light_status","monitor","t3"))
dim(a3o51f)
a3o51f$experiment<-rep("Free-run",length(a3o51f$time_step))
a3o51f$trik_position<-rep(3,length(a3o51f$time_step))
names(a3o51f)[7]<-c("counts")

###stitching
a3o51full<-rbind(a3o51,a3o51f)
a3o51full$uniqueID<-rep("a3o51",length(a3o51full$time_step))
head(a3o51full)
```

**10w15**
```{r 10w15}
a10w15<-subset(allm,monitor =="2" & time_step <41988 & time_step>34287, select=c("time_step","date","time","trik_status","light_status","monitor","t12"))  

a10w15$experiment<-rep("Entrainment",length(a10w15$time_step))

a10w15$trik_position<-rep(12,length(a10w15$time_step))

names(a10w15)[7]<-c("counts") 

###free run
a10w15f<-subset(allm,monitor =="5" & time_step <52917 & time_step>41988, select=c("time_step","date","time","trik_status","light_status","monitor","t10"))
dim(a10w15f)
a10w15f$experiment<-rep("Free-run",length(a10w15f$time_step))
a10w15f$trik_position<-rep(10,length(a10w15f$time_step))
names(a10w15f)[7]<-c("counts")

###stitching
a10w15full<-rbind(a10w15,a10w15f)
a10w15full$uniqueID<-rep("a10w15",length(a10w15full$time_step))
head(a10w15full)
```

**2b23**
```{r 2b23}
a2b23<-subset(allm,monitor =="2" & time_step <34272 & time_step>26049, select=c("time_step","date","time","trik_status","light_status","monitor","t16"))  

a2b23$experiment<-rep("Entrainment",length(a2b23$time_step))

a2b23$trik_position<-rep(16,length(a2b23$time_step))

names(a2b23)[7]<-c("counts") 

###free run
a2b23f<-subset(allm,monitor =="4" & time_step <52917 & time_step>34341, select=c("time_step","date","time","trik_status","light_status","monitor","t8"))
dim(a2b23f)
a2b23f$experiment<-rep("Free-run",length(a2b23f$time_step))
a2b23f$trik_position<-rep(8,length(a2b23f$time_step))
names(a2b23f)[7]<-c("counts")

###stitching
a2b23full<-rbind(a2b23,a2b23f)
a2b23full$uniqueID<-rep("a2b23",length(a2b23full$time_step))
head(a2b23full)
```

**2b26**
```{r 2b26}
a2b26<-subset(allm,monitor =="2" & time_step <34262 & time_step>26049, select=c("time_step","date","time","trik_status","light_status","monitor","t19"))  

a2b26$experiment<-rep("Entrainment",length(a2b26$time_step))

a2b26$trik_position<-rep(19,length(a2b26$time_step))

names(a2b26)[7]<-c("counts") 

###free run
a2b26f<-subset(allm,monitor =="4" & time_step <52917 & time_step>34253, select=c("time_step","date","time","trik_status","light_status","monitor","t5"))
dim(a2b26f)
a2b26f$experiment<-rep("Free-run",length(a2b26f$time_step))
a2b26f$trik_position<-rep(5,length(a2b26f$time_step))
names(a2b26f)[7]<-c("counts")

###stitching
a2b26full<-rbind(a2b26,a2b26f)
a2b26full$uniqueID<-rep("a2b26",length(a2b26full$time_step))
head(a2b26full)
```

**4o15**
```{r 4o15}
a4o15<-subset(allm,monitor =="2" & time_step <34905 & time_step>27493, select=c("time_step","date","time","trik_status","light_status","monitor","t29"))  

a4o15$experiment<-rep("Entrainment",length(a4o15$time_step))

a4o15$trik_position<-rep(29,length(a4o15$time_step))

names(a4o15)[7]<-c("counts") 

###free run
a4o15f<-subset(allm,monitor =="5" & time_step <52917 & time_step>34905, select=c("time_step","date","time","trik_status","light_status","monitor","t30"))
dim(a4o15f)
a4o15f$experiment<-rep("Free-run",length(a4o15f$time_step))
a4o15f$trik_position<-rep(30,length(a4o15f$time_step))
names(a4o15f)[7]<-c("counts")

###stitching
a4o15full<-rbind(a4o15,a4o15f)
a4o15full$uniqueID<-rep("a4o15",length(a4o15full$time_step))
head(a4o15full)
```

**5w63**
```{r 5w63}
a5w63<-subset(allm,monitor =="2" & time_step <34905 & time_step>27492, select=c("time_step","date","time","trik_status","light_status","monitor","t9"))  

a5w63$experiment<-rep("Entrainment",length(a5w63$time_step))

a5w63$trik_position<-rep(9,length(a5w63$time_step))

names(a5w63)[7]<-c("counts") 

###free run
a5w63f<-subset(allm,monitor =="5" & time_step <52917 & time_step>34905, select=c("time_step","date","time","trik_status","light_status","monitor","t16"))
dim(a5w63f)
a5w63f$experiment<-rep("Free-run",length(a5w63f$time_step))
a5w63f$trik_position<-rep(16,length(a5w63f$time_step))
names(a5w63f)[7]<-c("counts")

###stitching
a5w63full<-rbind(a5w63,a5w63f)
a5w63full$uniqueID<-rep("a5w63",length(a5w63full$time_step))
head(a5w63full)
```

**11w26**
```{r 11w26}
a11w26<-subset(allm,monitor =="2" & time_step <41988 & time_step>34264, select=c("time_step","date","time","trik_status","light_status","monitor","t25"))  

a11w26$experiment<-rep("Entrainment",length(a11w26$time_step))

a11w26$trik_position<-rep(25,length(a11w26$time_step))

names(a11w26)[7]<-c("counts") 

###free run
a11w26f<-subset(allm,monitor =="5" & time_step <52917 & time_step>41988, select=c("time_step","date","time","trik_status","light_status","monitor","t27"))
dim(a11w26f)
a11w26f$experiment<-rep("Free-run",length(a11w26f$time_step))
a11w26f$trik_position<-rep(27,length(a11w26f$time_step))
names(a11w26f)[7]<-c("counts")

###stitching
a11w26full<-rbind(a11w26,a11w26f)
a11w26full$uniqueID<-rep("a11w26",length(a11w26full$time_step))
head(a11w26full)
```

**10o62**
```{r 10o62}
a10o62<-subset(allm,monitor =="1" & time_step <41427 & time_step>37123, select=c("time_step","date","time","trik_status","light_status","monitor","t31"))  

a10o62$experiment<-rep("Entrainment",length(a10o62$time_step))

a10o62$trik_position<-rep(31,length(a10o62$time_step))

names(a10o62)[7]<-c("counts") 
```

**4o25**
```{r 4o25}
a4o25<-subset(allm,monitor =="2" & time_step <47187 & time_step>42151, select=c("time_step","date","time","trik_status","light_status","monitor","t3"))  

a4o25$experiment<-rep("Entrainment",length(a4o25$time_step))

a4o25$trik_position<-rep(3,length(a4o25$time_step))

names(a4o25)[7]<-c("counts") 

###free run
a4o25f<-subset(allm,monitor =="6" & time_step <52917 & time_step>47186, select=c("time_step","date","time","trik_status","light_status","monitor","t12"))
dim(a4o25f)
a4o25f$experiment<-rep("Free-run",length(a4o25f$time_step))
a4o25f$trik_position<-rep(12,length(a4o25f$time_step))
names(a4o25f)[7]<-c("counts")

###stitching
a4o25full<-rbind(a4o25,a4o25f)
a4o25full$uniqueID<-rep("a4o25",length(a4o25full$time_step))
head(a5b25full)
```

**h4w10**
```{r h4w10}
h4w10<-subset(allm,monitor =="2" & time_step <45272 & time_step>38649, select=c("time_step","date","time","trik_status","light_status","monitor","t9"))  

h4w10$experiment<-rep("Entrainment",length(h4w10$time_step))

h4w10$trik_position<-rep(9,length(h4w10$time_step))

names(h4w10)[7]<-c("counts") 

###free run
h4w10f<-subset(allm,monitor =="5" & time_step <52917 & time_step>45272, select=c("time_step","date","time","trik_status","light_status","monitor","t21"))
dim(h4w10f)
h4w10f$experiment<-rep("Free-run",length(h4w10f$time_step))
h4w10f$trik_position<-rep(21,length(h4w10f$time_step))
names(h4w10f)[7]<-c("counts")

###stitching
h4w10full<-rbind(h4w10,h4w10f)
h4w10full$uniqueID<-rep("h4w10",length(h4w10full$time_step))
head(h4w10full)
```

###h8w4 and h8o14 both have not exited the trikinetics###
**h8w4**
```{r h8w4}
h8w4<-subset(allm,monitor =="1" & time_step <52917 & time_step>34784, select=c("time_step","date","time","trik_status","light_status","monitor","t12"))
h8w4$experiment<-rep("Entrainment",length(h8w4$time_step))

h8w4$trik_position<-rep(12,length(h8w4$time_step))

names(h8w4)[7]<-c("counts") 
```

***h8o14***
```{r h8o14}
h8o14<-subset(allm,monitor =="1" & time_step <52917 & time_step>34784, select=c("time_step","date","time","trik_status","light_status","monitor","t6"))
h8o14$experiment<-rep("Entrainment",length(h8o14$time_step))

h8o14$trik_position<-rep(6,length(h8o14$time_step))

names(h8o14)[7]<-c("counts") 

h8o14$uniqueID<-rep("h8o14",nrow(h8o14))

```




### Adding all of the data together: 

```{r all.data stitch}

#all.data<-rbind(h4o4full,h2b25full,h5o22full,h8o14)   
all.data<-rbind(h4o4full,h2b25full,h5o22full,a12b6full,a5w73full,a4w66full,a10w12full,a10o49full,a12w40full,a12b43full,a13o28full,a13o11full,a5b25full,a10o73full,a3o51full,a10w15full,a2b23full,a2b26full,a4o15full,a5w63full,a11w26full,a5b25full,h4w10full)
all.data$uniqueID<-as.factor(all.data$uniqueID)
all.data$date<-as.factor(all.data$date)
names(all.data)
head(all.data)
dim(all.data)
#ggplot(all.data,aes(x=time_step,y=counts,colour=uniqueID))+geom_line()


ggplot(all.data,aes(x=time_step,y=counts))+geom_line()+facet_grid(uniqueID~.,scale="free")



###convert time to hours   

all.data$time2<-hour(hms(all.data$time))+minute(hms(all.data$time))/60

```

### 1 hour bin 
```{r}
bins60=c(paste0(rep(c(paste0(0,0:9),10:23), each=1),":", c("00"))[-1],"24:00")
all.data$bins60=cut(all.data$time2,breaks=seq(0,24,1),labels=bins60)

head(all.data)
#all.data2<-ddply(all.data,.(date,uniqueID,experiment,bins60),transform,time=seq(1,length(bins60),1))
nall.data<-ddply(all.data,.(uniqueID,date,experiment,bins60),summarize,counts=sum(counts))

nall.data2<-na.omit(ddply(nall.data,.(uniqueID,experiment),transform,order(date)))
nall.data3<-ddply(nall.data2,.(uniqueID),transform,time=seq(1,length(bins60),1))
#counts15[order(as.Date(counts15$date,format ="%Y-%m-%d")),]

dim(nall.data3)
head(nall.data3)

ggplot(nall.data3,aes(x=time,y=counts))+geom_line()+facet_grid(uniqueID~.,scales="free")#+geom_vline(xintercept=seq(0,900,24))#+geom_vline(xintercept=192)
```

### 15 min bins   

```{r}
bins15=c(paste0(rep(c(paste0(0,0:9),10:23), each=4),":", c("00",15,30,45))[-1],"24:00")
all.data$bins15=cut(all.data$time2,breaks=seq(0,24,.25),labels=bins15)

nall.data15<-ddply(all.data,.(uniqueID,date,experiment,bins15),summarize,counts=sum(counts))
nall.data15_2<-na.omit(ddply(nall.data15,.(uniqueID,experiment),transform,order(date)))
nall.data15_3<-ddply(nall.data15_2,.(uniqueID),transform,time=seq(1,length(bins15),1))
#write.csv(nall.data15_3,"2017-12-08_15min_bins_subset_behavior.csv")
ggplot(nall.data15_3,aes(x=time,y=counts))+geom_line()+facet_grid(uniqueID~.,scales="free")

###adding time in days
nall.data15_3$days<-nall.data15_3$time/4/24
ggplot(nall.data15_3,aes(x=days,y=counts))+geom_line()+facet_grid(uniqueID~.,scales="free")
```

###6 min bins

```{r}
bins6=c(paste0(rep(c(paste0(0,0:9),10:23), each=10),":", c("00","06",12,18,24,30,36,42,48,54))[-1],"24:00")

all.data$bins06=cut(all.data$time2,breaks=seq(0,24,.1),labels=bins6)

nall.data06<-ddply(all.data,.(uniqueID,date,experiment,bins06),summarize,counts=sum(counts))
nall.data06_2<-na.omit(ddply(nall.data06,.(uniqueID,experiment),transform,order(date)))
nall.data06_3<-ddply(nall.data06_2,.(uniqueID),transform,time=seq(1,length(bins06),1))
#write.csv(nall.data06_3,"2017-12-08_15min_bins_subset_behavior.csv")
ggplot(nall.data06_3,aes(x=time,y=counts))+geom_line()+facet_grid(uniqueID~.,scales="free")

###adding time in days
nall.data06_3$days<-nall.data06_3$time/10/24
ggplot(nall.data06_3,aes(x=days,y=counts))+geom_line()+facet_grid(uniqueID~.,scales="free")
```

# scaling up   

### periodogram   

```{r}
pmod1<-periodogram(nall.data15_3$counts,plot=TRUE,xlim=c(0,.015))
pmod1.vals<-data.frame(freq=pmod1$freq,pmod1$spec)
pmod1.order<-pmod1.vals[order(-pmod1.vals$pmod1.spec),]
top10<-head(pmod1.order,10);top10

1/top10$freq/4

1/top10$freq/4/24

##making a function to grab top period values 
per.fun<-function(ts=nall.data15_3$counts){
m<-data.frame(freq=periodogram(ts,plot=FALSE)$freq,spec=periodogram(ts,plot=FALSE)$spec)
return(1/(m[order(m$spec,decreasing=TRUE),][1:3,1])/4/24)
}

### using function in ddply 
## for each unique ID
test<-ddply(nall.data15_3,.(uniqueID),function(sub) per.fun(sub$counts))
#summary(nall.data15_3$uniqueID)
test

### for each unique ID and experiment  
test2<-ddply(nall.data15_3,.(uniqueID,experiment),function(sub) per.fun(sub$counts))
test2

#tt<-subset(nall.data15_3,uniqueID=="a11w26")
#per.fun(ts=tt$counts)
#periodogram(tt$counts)


### ok lets compare comparable rhythms  
test2$min<-as.numeric(apply(test2,1,min))
test2$max<-as.numeric(apply(test2[,-1:-2],1,max))
test2$host<-ifelse(substr(test2$uniqueID,1,1)=="a","apple","haw")
test2

summary(aov(min~host*experiment,data=test2))
summary(aov(log10(max)~host*experiment,data=test2))

#ggplot(test2,aes(x=host,y=min))+geom_boxplot()
#ggplot(test2,aes(x=host,y=max))+geom_boxplot()

a<-ggplot(test2,aes(x=experiment,y=log10(min)))+geom_boxplot()

b<-ggplot(test2,aes(x=experiment,y=log(min),group=uniqueID,colour=host))+geom_point()+geom_line()

grid.arrange(nrow=2,a,b)
```

### spectral analysis   

```{r}
sa.an<-function(ts=counts15$counts15){
  sa1<-spectrum(ts,method=c("pgram","ar"),plot=FALSE,demean=TRUE,detrend=TRUE,tape=.2)
  spx<-sa1$freq
  spy<-2*sa1$spec
  pw<-data.frame(spx,spy)
  cc<-head(pw[order(pw$spy,decreasing=TRUE),],4)
  #return(1/cc[,1]/4)  ## hours 
  return(1/cc[,1]/4/24)  ## days 
  
}
sa.an<-function(ts=counts15$counts15){
  sa1<-spectrum(ts,method=c("pgram","ar"),plot=FALSE,demean=TRUE,detrend=TRUE,tape=.2)
  spx<-sa1$freq
  spy<-2*sa1$spec
  pw<-data.frame(spx,spy)
  cc1<-pw[order(pw$spy,decreasing=TRUE),]
  cc2<-subset(cc1,spx<0.05)
  cc2$density<-density(cc2$spy,n=length(cc2$spy))$y
  cc2$t<-density(cc2$spy,n=length(cc2$spy))$x
  #cc<-findpeaks(cc2[,3],minpeakheight=1E-6)
  cc<-findpeaks(cc2[,1])
  cc2[order(cc2$density,decreasing=TRUE),]
  out<-1/cc2[cc[,2],][,1]/4
  #out<-1/cc2[cc[,2],][,1]/4/24
  return(out[1:4])
  ## hours 
  
  
}

sa.an()

### for each unique ID and experiment  
test10<-ddply(nall.data15_3,.(uniqueID,experiment),function(sub) sa.an(sub$counts))
test10


###match data with eclosion data 

### have to adjust uniqueID dataframe in test10 to match orig.dat
test10$uniqueID<-as.factor(ifelse(substr(test10$uniqueID,1,1)=="a",substr(test10$uniqueID,2,10),substr(test10$uniqueID,1,10)))

##orig.dat
orig.dat<-read.csv("../Data/2017-11-17_subset_host_comparison_trik_data_extract.csv")

test10.merg<-inner_join(test10,orig.dat,by="uniqueID")
head(test10.merg)
str(test10.merg)


##taking circadian value 

test10.merg$circ_time<-apply(test10.merg[,3:7],1,function(x){mean(subset(x,x<35 & x > 20))})

test10.merg$ult_time<-apply(test10.merg[,3:7],1,function(x){mean(subset(x,x<20))})
test10.merg$day_time<-apply(test10.merg[,3:7],1,function(x){mean(subset(x,x>48))})




ggplot(test10.merg,aes(x=organism,y=V1))+geom_boxplot()

ab<-ggplot(test10.merg,aes(x=organism,y=circ_time))+geom_boxplot()
cd<-ggplot(test10.merg,aes(x=organism,y=ult_time))+geom_boxplot()
ggplot(test10.merg,aes(x=organism,y=day_time))+geom_boxplot()

grid.arrange(cd,ab,ncol=2)

#ggplot(test10.merg,aes(y=eclosion_days,x=V1,colour=organism))+geom_point()+stat_smooth(method="lm")

ggplot(test10.merg,aes(y=eclosion_days,x=circ_time,colour=paste(experiment,organism)))+geom_point()+stat_smooth(method="lm")+ylim(0,100)

ggplot(test10.merg,aes(y=eclosion_days,x=V1,colour=paste(experiment,organism)))+geom_point()+stat_smooth(method="lm")+ylim(0,100)

#knitr::kable(ddply(test10.merg,.(organism,experiment,Host),summarize,count=length(organism)))


ggplot(test10.merg,aes(y=eclosion_days,x=ult_time,colour=paste(experiment,organism)))+geom_point()+stat_smooth(method="lm")+ylim(0,100)

ggplot(test10.merg,aes(y=eclosion_days,x=day_time,colour=paste(experiment,organism)))+geom_point()+stat_smooth(method="lm")+ylim(0,100)



###stats
summary(aov(eclosion_days~organism+circ_time,data=test10.merg)) 
summary(aov(eclosion_days~organism+ult_time,data=test10.merg)) 
#summary(lm(eclosion_days~organism*circ_time*experiment,data=test10.merg)) 


summary(aov(eclosion_days~organism*V1,data=test10.merg)) 

summary(aov(circ_time~organism,data=test10.merg)) 


tt<-subset(test10.merg,organism!="wasp")
summary(aov(circ_time~Host,data=tt))

```

### trying continuous wavelet analysis 

```{r}
#, p2 = 8
cwa<-function(series=counts15$counts15,N=4){
  wave.out <- morlet(y1 = series,p2=8, dj = 0.1, siglvl = 0.95)
  wave.out$period <- wave.out$period/N
  wave.avg <- data.frame(power = apply(wave.out$Power, 2, mean), period = (wave.out$period))
  aa<-data.frame(findpeaks(wave.avg$power))
  line<-aa[order(aa,decreasing=TRUE),][1,2]
  return(wave.avg[line,][2])
  }

cwa()


###15 min
test3<-ddply(nall.data15_3,.(uniqueID,experiment),function(sub) cwa(sub$counts))
knitr::kable(test3)


ggplot(test3,aes(x=experiment,y=period,group=uniqueID,colour=uniqueID))+geom_point()+geom_line()

###6 min
test3.6<-ddply(nall.data06_3,.(uniqueID,experiment),function(sub) cwa(sub$counts,N=10))
knitr::kable(test3.6)

ggplot(test3.6,aes(x=experiment,y=period,group=uniqueID,colour=uniqueID))+geom_point()+geom_line()

ggplot(test3.6,aes(y=period,x=experiment))+geom_boxplot()

n<-subset(nall.data06_3,uniqueID=="a4o15"&experiment=="Entrainment")
nn<-morlet(n$counts)
wave.avg <- data.frame(power = apply(nn$Power, 2, mean), period = (nn$period))

######
cwa.1<-function(series=counts15$counts15,N=4){
  wave.out <- morlet(y1 = series, p2 = 8, dj = 0.1, siglvl = 0.95)
  wave.out$period <- wave.out$period/N
  x<-data.frame(power = apply(wave.out$Power, 2, mean), period = (wave.out$period))
  x$density<-density(x$power,n=length(x$power))$y
  return(x)
  #return(data.frame(findpeaks(wave.avg$power)))
  #line<-aa[order(aa,decreasing=TRUE),][1,2]
  #return(wave.avg[line,][2])
  }

plot(cwa.1()[,2:1],type="line")

tt3<-ddply(nall.data06_3,.(uniqueID,experiment),function(sub) cwa.1(sub$counts,N=10))

#ggplot(tt3,aes(x=period,y=power))+facet_wrap(uniqueID~experiment,scale="free")+geom_line()

ggplot(tt3,aes(x=period,y=density))+facet_wrap(uniqueID~experiment,scale="free")+geom_line()



###retaining all peaks   

allpeaks<-function(series=counts15$counts15,N=4){
  wave.out <- morlet(y1 = series, p2 = 8, dj = 0.1, siglvl = 0.95)
  wave.out$period <- wave.out$period/N
  wave.avg <- data.frame(power = apply(wave.out$Power, 2, mean), period = (wave.out$period))
  aa<-data.frame(findpeaks(wave.avg$power))[,1:2]
  aa$period<-wave.avg[aa$X2,2]
  names(aa)<-c("Power","time","period")
  return(aa)
}

allpeaks()

tt4<-ddply(nall.data06_3,.(uniqueID,experiment),function(sub) allpeaks(sub$counts,N=10))

tt5<-subset(tt4,Power>100)
tt5

tt5$hr<-as.numeric(substr(tt5$period,0,2))

ddply(tt5,.(uniqueID,experiment,hr),summarize,count=length(hr))

tt5$hr.trans<-ifelse(tt5$hr>20,"24",ifelse(tt5$hr <9,"4-8","9-13"))

ddply(tt5,.(experiment,hr.trans),summarize,count=length(hr.trans))

table(ddply(tt5,.(uniqueID,experiment,hr.trans),summarize,count=length(hr.trans)))

rhythm.num<-ddply(subset(tt5,experiment=="Free-run"),.(uniqueID),summarize,length=length(hr.trans))

rhythm.num$num<-ifelse(rhythm.num$length>1,"1","0")

rhythm.num$uniqueID<-as.factor(ifelse(substr(rhythm.num$uniqueID,1,1)=="a",substr(rhythm.num$uniqueID,2,10),substr(rhythm.num$uniqueID,1,10)))
b<-inner_join(rhythm.num,orig.dat,by="uniqueID")

ggplot(b,aes(x=num,y=eclosion_days))+geom_boxplot()

summary(aov(eclosion_days~as.factor(num),data=b))
```

### ok discrete wavelet analysis  

```{r}
## creating function for discrete wavelet analysis
### getting the circadian detail  
dwa<-function(time,levels=seq(5,7)){
  DJt_circ <- wavMRDSum(time,levels=levels ,keep.smooth=FALSE, keep.details=TRUE,reflect=TRUE,wavelet="s12",xform="modwt")
  IBL<-data.frame(findpeaks(DJt_circ))
  names(IBL)<-c("Height","mid_time","initial_time","final_time")
  #return(diff(IBL$mid_time)/4)
  return(data.frame(cbind(IBL[,-3:-4],bout=as.numeric(c("NA",diff(IBL$mid_time)))),act_length=IBL[,4]-IBL[,3]))
}
plot(dwa(time=counts15$counts15))
#dwa(time=counts15$counts15)

#plot(dwa(time=counts15$counts15)[,2]/4/24,dwa(time=counts15$counts15)[,1],type="l")

ll<-subset(nall.data15_3,experiment=="Free-run")
#ll<-subset(nall.data3,experiment=="Free-run")
#ll<-subset(nall.data06_3,experiment=="Free-run")

test4<-ddply(ll,.(uniqueID),function(sub) dwa(sub$counts,levels=c(6)))
#test4<-ddply(nall.data06_3,.(uniqueID),function(sub) dwa(sub$counts))
test4
test4$uniqueID<-as.factor(ifelse(substr(test4$uniqueID,1,1)=="a",substr(test4$uniqueID,2,10),substr(test4$uniqueID,1,10)))

#ggplot(test4.2,aes(x=bout,y=Height))+geom_point()+facet_wrap(~uniqueID,ncol=2)+stat_smooth()
sumtest4<-ddply(test4,.(uniqueID),summarize,bouts=mean(bout,na.rm=TRUE),sd=sd(bout,na.rm=TRUE))
sumtest4
rt<-inner_join(sumtest4,orig.dat,by="uniqueID")


ggplot(rt,aes(x=bouts/4,y=eclosion_days))+geom_point()+stat_smooth(method="lm")
#ggplot(rt,aes(x=sd,y=eclosion_days))+geom_point()+stat_smooth(method="lm")
summary(lm(eclosion_days~bouts,data=rt))

#plot(lm(eclosion_days~bouts,data=rt))
###############




test4.2<-ddply(nall.data15_3,.(uniqueID),function(sub) dwa(sub$counts,levels=seq(3:7)))
test4.2

ggplot(test4.2,aes(x=mid_time/4/24,y=bout))+geom_line()+facet_wrap(~uniqueID,ncol=2)#+geom_point(size=3)

```


# some prelim analyses just on h4o4full   

## plots  

 
```{r}
plot(h4o4full$counts,type="l")
h4o4full$date<-as.factor(h4o4full$date)
##adding time 
head(h4o4full)

h4o4full$time_hr<-hour(hms(h4o4full$time))+minute(hms(h4o4full$time))/60
#write.csv(h4o4full[order(h4o4full$date),],"test.csv")

h4o4full<-h4o4full[order(h4o4full$date),]

```

## bin by 6 minutes
```{r}
bins6=c(paste0(rep(c(paste0(0,0:9),10:23), each=10),":", c("00","06",12,18,24,30,36,42,48,54))[-1],"24:00")

h4o4full$bins6=cut(h4o4full$time_hr,breaks=seq(0,24,.1),labels=bins6)

```

## bin by 15 minutes 
```{r}
bins15=c(paste0(rep(c(paste0(0,0:9),10:23), each=4),":", c("00",15,30,45))[-1],"24:00")
h4o4full$bins15=cut(h4o4full$time_hr,breaks=seq(0,24,.25),labels=bins15)

```

## 30 min bins

```{r}
bins30=c(paste0(rep(c(paste0(0,0:9),10:23), each=2),":", c("00",30))[-1],"24:00")
length(bins30)
h4o4full$bins30=cut(h4o4full$time_hr,breaks=seq(0,24,.5),labels=bins30)

#h4o4full<-h4o4full[order(as.Date(h4o4full$date,format ="%Y/%m/%d")),]
```

### 1 hour bins  

```{r}
bins60=c(paste0(rep(c(paste0(0,0:9),10:23), each=1),":", c("00"))[-1],"24:00")
h4o4full$bins60=cut(h4o4full$time_hr,breaks=seq(0,24,1),labels=bins60)
```

### summing based on bins   

```{r}

## 60 min or 1 hr bins

counts60<-ddply(h4o4full,.(date,experiment,bins60),summarize,counts60=sum(counts))

#lets try 30 min bins  

counts30<-ddply(h4o4full,.(date,experiment,bins30),summarize,counts30=sum(counts))

## 15 min bins
counts15<-na.omit(ddply(h4o4full,.(date,experiment,bins15),summarize,counts15=sum(counts)))

#counts15.1<-counts15[order(as.Date(counts15$date,format ="%Y/%m/%d")),]
counts15<-counts15[-1:-18,]

counts15$time<-seq(0.25,801.25,.25)#[-3224]
counts15$day<-counts15$time/24
dim(h4o4full)
dim(counts15)

plot(counts15$day,counts15$counts15,type="l")


### 6 min bins

counts06<-ddply(h4o4full,.(date,experiment,bins6),summarize,counts06=sum(counts))

dim(h4o4full)
dim(counts06)

plot(counts06$counts06,type="l")

```


### Spectral analysis 


```{r}
spec.ar(counts15$counts15)
spec.ar(counts15$counts15,method="burg")
spec.ar(counts15$counts15,method="yule-walker")
val<-spec.ar(counts15$counts15,method="burg")

#detrending  

spec.ar(ar(counts15$counts15),plot=TRUE)
#spec.ar(ar(counts15$counts15),method="yule-walker",detrend=TRUE,demean = TRUE,log="no",spans=9,taper=0)
spec.pgram(counts15$counts15,method="yule-walker",detrend=TRUE,demean = TRUE,log="no",spans=c(6,6),taper=0,xlim=c(0,.02))

h<-spec.ar(ar(counts15$counts15),method="yule-walker")
str(h)
h<-data.frame(freq=h$freq,spec=h$spec)
tt<-h[order(h$spec,decreasing=TRUE),]
head(tt,20)

1/head(tt,20)[,1]/4/24


acf(counts15$counts15,type="correlation")
pacf(counts15$counts15)


library(psd)

xx<-pspectrum(counts15$counts15,AR=TRUE,niter=10)
plot(xx)
```

### Single spectrum analysis   
https://cran.r-project.org/web/packages/Rssa/Rssa.pdf

```{r}

#install.packages("Rssa")
library(Rssa)

s<-ssa(counts15$counts15)
summary(s)

r<-reconstruct(s)
plot(r)

ss<-decompose(s)


```


### Periodogram ; Discrete fourier transform  
http://www.di.fc.ul.pt/~jpn/r/fourier/fourier.html#fourier-transform    

Fourier transform explained  https://www.r-bloggers.com/the-fourier-transform-explained-in-one-sentence/
```{r}

##15 min bins
sa1<-spectrum(counts15$counts15,method=c("pgram","ar"))
sa1<-spectrum(counts15$counts15,method=c("pgram","ar"),demean=TRUE,detrend=TRUE,tape=.2)

spx<-sa1$freq
spy<-2*sa1$spec

pw<-data.frame(spx,spy)


plot(spy~spx,xlab="freq",ylab="spectral density",type="l",xlim=c(0,.5))
plot(spy~spx,xlab="freq",ylab="spectral density",type="l",xlim=c(0,.05))
#1/0.045

cc<-head(pw[order(pw$spy,decreasing=TRUE),],4)
1/cc[,1]/4  ## hours 
1/cc[,1]/4/24  ## days 


################################################################################
##smoothing 

#spec.pgram automatically detrends
################################################################################

raw.spec1<-spec.pgram(counts15$counts15,taper=0.1,kernel=c(5),spans=5,detrend=TRUE,log="no",demean = TRUE)
plot(raw.spec1,log="no",xlim=c(0,.02))
#(1/.01)/4/24

rs2<-data.frame(f=raw.spec1$freq,s=raw.spec1$spec)
rs3<-head(rs2[order(rs2$s,decreasing=TRUE),],10)
## the units of frquency are #cycles/unit time; so in this case #cycles/0.25hr
#multiply by .25 or divide 4 to get #cycles/hour 
1/rs3[,1]/4 # hours
1/rs3[,1]/4/24 # days

#1hr bin
raw.spec2<-spec.pgram(counts60$counts60,taper=0.1,,kernel=c(5),spans=5,detrend=TRUE,demean=TRUE)
#plot(raw.spec2,xlim=c(0,.1))

rs21<-data.frame(f=raw.spec2$freq,s=raw.spec2$spec)
rs31<-head(rs2[order(rs21$s,decreasing=TRUE),],10)
1/rs31[,1]/24


#plots for 1 hr
#plot(seq(1,nrow(counts60),1)/24,counts60$counts60,type="l")
#plot(seq(1,nrow(counts60),1)/24,counts60$counts60,type="l",xlim=c(0,8))
#plot(seq(1,nrow(counts60),1)/24,counts60$counts60,type="l",xlim=c(8,35))

################################################################################
################################################################################
#https://anomaly.io/detect-seasonality-using-fourier-transform-r/
################################################################################

library(TSA)

#60 min 
pp<-periodogram(counts60$counts60,plot=TRUE,xlim=c(0,.1))
dd00<-data.frame(freq=pp$freq,pp$spec)
order00<-dd00[order(-dd00$pp.spec),]
top00<-head(order00,10);top00

1/top00$freq

1/top00$freq/24



#30 min
p0<-periodogram(counts30$counts30,plot=TRUE,xlim=c(0,.1))
p0

dd0<-data.frame(freq=p0$freq,p0$spec)
order0<-dd0[order(-dd0$p0.spec),]
top0<-head(order0,10);top0
1/top0$freq/2
1/top0$freq/2/24


# 15 min 
p<-periodogram(counts15$counts15,plot=TRUE,xlim=c(0,.04))
p

dd<-data.frame(freq=p$freq,p$spec)
order<-dd[order(-dd$p.spec),]
top2<-head(order,10);top2

#1/top2$freq

(1/top2$freq)/4 # to get hours 
(1/top2$freq)/4/24 # days





### 6 min
p2<-periodogram(counts06$counts06,plot=TRUE,xlim=c(0,.02))
dd2<-data.frame(freq=p2$freq,p2$spec)
order2<-dd2[order(-dd2$p2.spec),]
top10<-head(order2,10);top2
#1/top10$freq

(1/top10$freq)/10 # to get hours 
((1/top10$freq)/10)/24 # days


plot(p2)
```


### Continuous wavelet analysis 
http://rstudio-pubs-static.s3.amazonaws.com/9428_1197bd003ebd43c49b429f22ea4f36e5.html     

finding the dominant period   

```{r}
library(dplR)

# try 1 hour interval

wave.out00 <- morlet(y1 = counts60$counts60, p2 = 8, dj = 0.1, siglvl = 0.95)
#wave.out00$period <- wave.out00$period
wavelet.plot(wave.out00)

wave.avg <- data.frame(power = apply(wave.out00$Power, 2, mean), period = (wave.out00$period))

findpeaks(wave.avg$power)
wave.avg[findpeaks(wave.avg$power)[3,2],] #period in hours 
### 30 min
wave.out0 <- morlet(y1 = counts30$counts30, p2 = 8, dj = 0.1, siglvl = 0.95)
#wave.out0$period <- wave.out0$period/2/24
wave.out0$period <- wave.out0$period/2
wavelet.plot(wave.out0)

wave.avg <- data.frame(power = apply(wave.out0$Power, 2, mean), period = (wave.out0$period))

findpeaks(wave.avg$power)
wave.avg[findpeaks(wave.avg$power)[2,2],] #period in hours 

##15 min
wave.out <- morlet(y1 = counts15$counts15,counts15$day, p2 = 8, dj = 0.1, siglvl = 0.95)
#wave.out$period <- wave.out$period/4/24
wave.out$period <- wave.out$period/4

#levs <- quantile(wave.out$Power, c(0, 0.25, 0.5, 0.75, 0.95, 1))
#wavelet.plot(wave.out, wavelet.levels = levs)
wavelet.plot(wave.out)


wave.avg <- data.frame(power = apply(wave.out$Power, 2, mean), period = (wave.out$period))

findpeaks(wave.avg$power)
#aa<-data.frame(findpeaks(wave.avg$power))
#aa[order(aa,decreasing=TRUE),][1,2]

wave.avg[findpeaks(wave.avg$power)[2,2],][2] #period in hours 

plot(wave.avg$period, wave.avg$power, type = "l")
abline(v=wave.avg[findpeaks(wave.avg$power)[2,2],])

###24 hour period slice 
#need to take mean from 66-67 column ( 23.372-25.052) = 24.21 
p24<-apply(wave.out$Power[,66:67],1,mean)
plot(wave.out$x/4/24,p24)
findpeaks(p24)



##6 min
wave.out2 <- morlet(y1 = counts06$counts06, p2 = 8, dj = 0.1, siglvl = 0.95)
wave.out2$period <- wave.out2$period/10

levs <- quantile(wave.out$Power, c(0, 0.25, 0.5, 0.75, 0.95, 1))
#wavelet.plot(wave.out, wavelet.levels = levs)
wavelet.plot(wave.out2)


wave.avg2 <- data.frame(power = apply(wave.out2$Power, 2, mean), period = (wave.out2$period))

findpeaks(wave.avg2$power)
wave.avg2[findpeaks(wave.avg2$power)[3,2],] 

plot(wave.avg2$period, wave.avg2$power, type = "l")
abline(v=wave.avg2[findpeaks(wave.avg2$power)[3,2],])

#4 hour rhythms
plot(wave.out2$x/10/24,wave.out2$Power[,54])
findpeaks(wave.out2$Power[,54])

#8 hour rhythms
plot(wave.out2$x/10/24,wave.out2$Power[,64])
findpeaks(wave.out2$Power[,64])



###just free run data 
free<-subset(counts15,experiment=="Free-run")
wave.out3 <- morlet(y1 = free$counts15,x1=free$day, p2 = 8, dj = 0.1, siglvl = 0.95)
wave.out3$period <- wave.out3$period/4
wavelet.plot(wave.out3)
#plot(free$day,free$counts15,type="l")


```

```{r}

op<-wavCWT(counts15$counts15)
plot(op)
str(op)
wavCWTPeaks(wavCWTTree(op))
```

### Discrete wavelet analysis   


```{r}
#calculating period levels for each time step   
detail.level<-seq(1,9)

##.1 hr 
2^(detail.level)*.1

##.25 hr 
2^(detail.level)*.25

##.5 hr 
2^(detail.level)*.5

##1 hr 
2^(detail.level)*1
(2^(detail.level)*1)/24 # days

##2 hr 
2^(detail.level)*2
(2^(detail.level)*2)/24 # days
```

```{r}
library(wmtsa)
Jcirc <- floor(log2(round(24/.25)))
floor(log2(round(24)))
#Jcirc <- floor(log2(round(18/.25)))

#all details
#DJt_circ <- wavMRDSum(counts15$counts15,levels=seq(1,8),keep.smooth=FALSE, keep.details=TRUE,reflect=TRUE,wavelet="s12",xform="modwt")

#circadian detail
##15 min
DJt_circ <- wavMRDSum(counts15$counts15,levels=Jcirc ,keep.smooth=FALSE, keep.details=TRUE,reflect=TRUE,wavelet="s12",xform="modwt")

##60 min 
#DJt_circ <- wavMRDSum(counts60$counts60,levels=6 ,keep.smooth=FALSE, keep.details=TRUE,reflect=TRUE,wavelet="s12",xform="modwt")


#highest detail
#DJt_circ <- wavMRDSum(counts15$counts15,levels=8 ,keep.smooth=FALSE, keep.details=TRUE,reflect=TRUE,wavelet="s12",xform="modwt")

DJt_circ

plot(DJt_circ,col=rgb(0,0,.5,.5),type="l")
#points(counts15$day,counts15$counts15,type="l",col=rgb(.5,0,0,.5))

library(pracma)
IBL<-data.frame(findpeaks(DJt_circ))
names(IBL)<-c("Height","mid_time","initial_time","final_time")
IBL
diff(IBL$mid_time)/4
#diff(IBL$mid_time)/24
#hist(diff(IBL$mid_time)/4/24)
hist(diff(IBL$mid_time)/4)
#hist(diff(IBL$mid_time))



plot(IBL$mid_time[-10]/4/24,diff(IBL$mid_time)/4/24)
lines(IBL$mid_time[-10]/4/24,diff(IBL$mid_time)/4/24)
#y=diff(IBL$mid_time)/4/24
#x=IBL$mid_time[-10]/4/24
#lines(loess(y~x,span=1))
dd<-data.frame(int=diff(IBL$mid_time)/4,time=IBL$mid_time[-36]/4/24)
dd$lighton=as.numeric(substr(dd$time,1,2))+6/24
head(dd)
dd$zeit<-dd$time-dd$lighton
plot(dd$time,dd$zeit)
lines(dd$time,dd$zeit)



#par(mfrow=c(2,1))
#par(mar=c(1,5,1,5))
#plot(IBL$mid_time[-31],diff(IBL$mid_time),xlab="time (.25 hr)",ylab="interval in .25 hr")
#plot(IBL$mid_time[-5],diff(IBL$mid_time),xlab="time (.25 hr)",ylab="interval in .25 hr")
plot(IBL$mid_time[-36]/4/24,diff(IBL$mid_time)/4,xlab="time (day)",ylab="interval in hr",las=1)
#lines(loess(diff(IBL$mid_time)~IBL$mid_time[-35],span=1))
s<-(IBL$mid_time[-36]/4/24)
lines(loess(diff(IBL$mid_time)/4~s,span=1))
abline(v=8,lty="dotdash")

#mean(diff(IBL$mid_time)/4)

# fitting polynomial function? 
model<-lm(diff(IBL$mid_time)/4~s+I(s^2)+I(s^3))
summary(model)

#plot(counts15$counts15,type="l")
#abline(v=seq(0,4000,96),lty="dotdash")
#points(IBL$mid_time,IBL$Height,pch=20,col="red")
#par(mfrow=c(1,1))
#par(mar=c(5,5,5,5))

plot(IBL$mid_time/4/24,IBL$Height,pch=20,col="red",type="l")
points(IBL$mid_time/4/24,IBL$Height,pch=20,col="red",cex=3)


du<-data.frame(bi=diff(IBL$mid_time)/4,height=IBL$Height[-1])
plot(du,pch=20,cex=3)
abline(summary(lm(height~bi,data=du)))
#summary(lm(height~I(bi^2),data=du))

#summary(lm(height~bi+I(bi^2),data=du))
```



### Session Information

```{r session info}
sessionInfo()
```
